Correlations between species co-occurrence and phylogenetic distances

Calculates measures of community phylogenetic structure (correlation between co-occurrence and phylogenetic distance) to patterns expected under various null models

Usage

comm.phylo.cor(
  samp,
  phylo,
  metric = c("cij", "checkerboard", "jaccard", "doij"),
  null.model = c("sample.taxa.labels", "pool.taxa.labels", "frequency", "richness",
    "independentswap", "trialswap"),
  runs = 999,
  ...
)

Arguments

samp

Community data matrix

phylo

Phylogenetic tree

metric

Metric of co-occurrence to use (see species.dist)

null.model

Null model to use (see Details section for description)

runs

Number of runs (randomizations)

...

Additional arguments to randomizeMatrix

Value

A list with elements:

obs.corr

Observed co-occurrence/phylogenetic distance correlation

obs.corr.p

P-value of observed correlation (standard P-value for correlation coefficient, not based on comparison with randomizations)

obs.rank

Rank of observed correlation vs. random

runs

Number of runs (randomizations)

obs.rand.p

P-value of observed correlation vs. randomizations (= obs.rank / (runs + 1))

random.corrs

A vector of random correlation calculated for each run

Details

Currently implemented null models (arguments to null.model):

sample.taxa.labels

Shuffle phylogeny tip labels (only within set of taxa present in community data)

pool.taxa.labels

Shuffle phylogeny tip labels (across all taxa included in phylogenetic tree)

frequency

Randomize community data matrix abundances within species (maintains species occurence frequency)

richness

Randomize community data matrix abundances within samples (maintains sample species richness)

independentswap

Randomize community data matrix maintaining species occurrence frequency and site richnessing using independent swap

trialswap

Randomize community data matrix maintaining species occurrence frequency and site richnessing using trial swap

References

Cavender-Bares J., D.A. Ackerly, D. Baum and F.A. Bazzaz. 2004. Phylogenetic overdispersion in Floridian oak communities, American Naturalist, 163(6):823-843.

See also

Author

Steven Kembel steve.kembel@gmail.com

Examples

data(phylocom)
comm.phylo.cor(phylocom$sample, phylocom$phylo, metric="cij",null.model="sample.taxa.labels")
#> $obs.corr
#> [1] -0.390639
#> 
#> $obs.corr.p
#> [1] 2.240087e-12
#> 
#> $obs.rank
#> [1] 1
#> 
#> $runs
#> [1] 999
#> 
#> $obs.rand.p
#> [1] 0.001
#> 
#> $random.corrs
#>   [1] -8.729103e-02  4.765843e-02 -4.653706e-02  7.351414e-02  4.871573e-02
#>   [6] -2.625619e-02  8.475992e-02 -1.385699e-02  1.260906e-01  2.795427e-02
#>  [11] -3.875151e-02  1.632401e-02  1.834249e-02 -1.808618e-02  5.765469e-02
#>  [16] -4.533559e-03  2.929992e-02  7.385055e-03  8.730705e-03  4.064183e-02
#>  [21] -1.106957e-02  1.747743e-02 -4.711377e-02 -1.193463e-02 -1.052651e-01
#>  [26] -8.517643e-02  5.366580e-03  4.309284e-03  2.535909e-02 -4.245205e-03
#>  [31] -2.385325e-02  1.209483e-02  9.302606e-02  6.707424e-02  6.053823e-02
#>  [36]  6.438294e-02 -2.039300e-02 -8.762744e-03 -1.000747e-01 -8.094725e-02
#>  [41]  2.382121e-02 -7.393065e-02  1.286858e-01  8.302980e-02  2.334062e-02
#>  [46]  3.444223e-03  3.977677e-02 -5.970520e-02 -7.758312e-02 -1.030063e-02
#>  [51]  1.810219e-03  9.610183e-02 -1.991241e-02 -2.837079e-02  9.023864e-02
#>  [56] -5.143907e-02  5.746245e-02 -3.288832e-02  3.381746e-02 -2.885137e-02
#>  [61] -2.414160e-02 -1.218935e-01 -7.018205e-02  2.209109e-02 -2.837079e-02
#>  [66] -8.210066e-02 -5.364978e-02  1.084530e-02 -4.115446e-02  3.151063e-02
#>  [71]  8.206862e-02  1.346451e-01 -8.238902e-02 -5.884014e-02 -3.932822e-02
#>  [76] -3.894375e-02 -7.705448e-03 -1.040155e-01  6.572859e-02  1.065306e-02
#>  [81] -5.302501e-03  6.024987e-02  1.295989e-02  4.910020e-02  5.159927e-02
#>  [86] -1.731723e-02 -5.076624e-02  1.521866e-03  3.948842e-02 -4.153893e-02
#>  [91] -1.059379e-01 -2.769796e-02  1.055694e-02  3.251987e-03 -1.019971e-01
#>  [96] -1.779782e-02  1.776578e-02 -5.393813e-02  2.151438e-02  1.173439e-01
#> [101]  1.978426e-02 -1.588027e-01 -1.625994e-02 -6.864416e-02 -3.298444e-02
#> [106]  5.467504e-02 -4.615259e-02  1.065787e-01 -7.849624e-04  1.425748e-03
#> [111] -1.491429e-02  2.016873e-02 -5.489931e-02  1.021573e-01  9.403530e-03
#> [116] -2.250760e-02  1.469001e-02  3.843112e-02  9.307412e-03 -1.155497e-01
#> [121]  7.985791e-02  1.805414e-02  1.151812e-02  4.169913e-02 -1.366475e-02
#> [126]  3.602817e-02  6.899660e-02  4.900408e-02  1.171036e-02  4.958079e-02
#> [131]  6.476741e-02  3.400970e-02  3.304852e-02  7.601320e-02 -1.195866e-01
#> [136]  5.813528e-02  7.226460e-02 -5.643720e-02  1.497836e-02  4.419819e-02
#> [141] -5.662943e-02 -1.693276e-02  4.073795e-02 -3.572380e-03  1.295508e-01
#> [146] -2.356489e-02  1.661237e-02 -9.940188e-02  4.044959e-02  1.670849e-02
#> [151] -1.962406e-02 -3.173491e-02 -6.806746e-02  9.033476e-02  6.409459e-02
#> [156] -6.441498e-02  6.467129e-02 -1.263149e-01  1.507448e-02  3.429805e-02
#> [161] -2.337266e-02 -1.076680e-01 -4.596035e-02  4.227583e-02 -5.230413e-02
#> [166] -1.808618e-02 -6.071444e-03 -4.509529e-02  2.237944e-02 -8.810803e-04
#> [171]  3.026110e-02 -4.471082e-02 -7.796760e-02  4.362148e-02  6.034599e-02
#> [176]  4.189136e-02 -4.533559e-03 -6.403051e-02  7.312966e-02  3.828694e-03
#> [181]  2.901157e-02  1.295989e-02  2.728145e-02  4.506325e-02 -5.912850e-02
#> [186]  7.274519e-02  3.795053e-02  2.194691e-03  2.314838e-02 -1.044000e-01
#> [191]  1.286377e-02  1.074918e-02  6.467129e-02 -1.231910e-02  2.170661e-02
#> [196] -1.364072e-01 -6.749075e-02  3.343299e-02  4.612055e-02 -2.231536e-02
#> [201]  3.804665e-02 -9.248139e-02  4.093018e-02 -5.783091e-03  9.595765e-03
#> [206]  3.439417e-02  3.266405e-02 -9.603775e-02  2.853098e-02 -1.760559e-02
#> [211]  4.506325e-02 -9.584552e-02  5.861587e-02 -3.029314e-02  4.737008e-02
#> [216]  3.775830e-02  3.276017e-02  3.795053e-02  2.603192e-02 -5.393813e-02
#> [221] -3.668498e-03 -6.681793e-02  5.515563e-02  6.966942e-02  2.795427e-02
#> [226]  4.823514e-02  6.890048e-02  9.754360e-02 -6.095474e-02 -2.942808e-02
#> [231] -5.903238e-02 -4.951671e-02 -6.787522e-02 -2.269983e-02 -1.351577e-01
#> [236] -2.500666e-02 -9.565328e-02 -5.398619e-03 -6.263680e-03  1.521866e-03
#> [241] -8.873279e-02 -7.335394e-02 -2.673678e-02  3.045334e-02 -7.897683e-03
#> [246]  5.438668e-02  9.302606e-02 -5.124683e-02  7.255296e-02 -1.193463e-02
#> [251] -9.200080e-02  1.030223e-01 -1.125700e-01 -1.379451e-01  1.038874e-01
#> [256]  4.693755e-03  3.516311e-02 -3.010091e-02  6.486353e-02 -1.501040e-02
#> [261]  1.920755e-02  4.333313e-02 -1.597158e-02 -6.508780e-02  3.679712e-02
#> [266]  4.535161e-02 -5.115072e-02 -9.738340e-02 -7.191217e-02 -5.624496e-02
#> [271]  1.718908e-02 -8.277349e-02  4.501519e-03  1.401719e-02 -5.206384e-03
#> [276]  3.670100e-02  4.016124e-02 -4.365352e-02  3.093393e-02 -6.739463e-02
#> [281]  5.366580e-03  4.631278e-02 -3.932822e-02  6.707424e-02 -2.404548e-02
#> [286]  2.949216e-02 -4.471082e-02  1.113365e-02  9.091147e-02  9.115176e-03
#> [291] -8.089919e-03  2.833875e-02  6.668977e-02  9.499647e-03 -2.740961e-02
#> [296]  3.304852e-02 -6.777910e-02 -1.098787e-01 -1.054573e-01  2.194691e-03
#> [301]  5.400221e-02 -4.644094e-02 -1.145404e-02  5.179150e-02  3.948842e-02
#> [306]  4.958079e-02  1.145565e-01  7.159178e-02 -1.748544e-01 -1.326587e-01
#> [311]  5.650128e-02  7.351414e-02  3.746994e-02  6.197999e-02  7.860838e-02
#> [316] -1.731723e-02 -3.865540e-02 -6.470333e-02  3.285628e-02 -5.355366e-02
#> [321]  6.755483e-02 -1.958081e-01 -7.047040e-02  8.360651e-02  3.843112e-02
#> [326] -3.663692e-02  8.697063e-02  2.612804e-02  2.545521e-02  2.314838e-02
#> [331]  6.640141e-02 -9.276974e-02  2.103379e-02 -3.875151e-02  4.506325e-02
#> [336] -1.777379e-01  5.207985e-02  4.986914e-02  6.005764e-02 -4.509529e-02
#> [341] -1.068510e-02 -4.596035e-02  3.823888e-02  3.189511e-02  6.024987e-02
#> [346]  1.031184e-01  5.015750e-02 -4.653706e-02  3.804665e-02  6.947719e-02
#> [351] -5.288084e-02 -7.633359e-02  3.093393e-02 -4.004910e-04  1.497836e-02
#> [356] -2.022480e-01 -1.010839e-02 -1.501040e-02 -5.076624e-02  1.766966e-02
#> [361]  5.534786e-02  2.391732e-02 -4.720988e-02  7.178401e-02  3.333687e-02
#> [366]  4.217972e-02  1.622790e-02 -4.437441e-03 -2.635231e-02  5.457892e-02
#> [371]  5.688575e-02 -6.508780e-02  2.372509e-02 -1.539488e-02 -5.182354e-02
#> [376] -2.193089e-02  7.385055e-03 -1.433758e-02 -1.193463e-02 -2.058524e-02
#> [381] -6.210815e-02  5.246433e-02  2.410956e-02  3.785441e-02  6.135523e-03
#> [386] -1.058898e-02 -7.897683e-03  2.295615e-02  4.544772e-02  9.821642e-02
#> [391] -2.145030e-02 -5.384202e-02  2.353285e-02 -1.847065e-02  5.943287e-03
#> [396]  8.898911e-02  4.179524e-02  3.689323e-02 -2.491054e-02  6.616112e-03
#> [401] -1.116569e-02  2.756980e-02 -7.095099e-02  1.055694e-02 -3.077373e-02
#> [406]  8.091521e-02  1.257542e-02  1.401719e-02  3.112616e-02  8.793181e-02
#> [411]  8.985417e-02  6.851601e-02  6.918883e-02 -2.707319e-03  1.034068e-01
#> [416]  8.860464e-02 -1.060340e-01  1.003310e-01 -1.314091e-01  1.459389e-02
#> [421]  1.603566e-02  1.997649e-02  3.429805e-02 -4.298070e-02  1.002349e-01
#> [426]  7.495590e-02 -7.162382e-02 -4.057775e-02 -3.519515e-02 -4.096222e-02
#> [431] -3.625245e-02  3.785441e-02  3.054946e-02 -1.568323e-02  8.331816e-02
#> [436] -6.263680e-03  3.737382e-02  2.535909e-02 -1.075719e-01  3.026110e-02
#> [441] -2.673678e-02 -1.138196e-01  4.746620e-02 -4.019328e-02  1.843861e-02
#> [446]  3.420193e-02  6.957330e-02 -5.576437e-02 -4.826718e-02 -4.437441e-03
#> [451] -4.009716e-02  1.334436e-02 -8.392690e-02  1.233512e-03  4.631278e-02
#> [456] -1.529876e-02  2.372509e-02  4.064183e-02 -5.028565e-02 -4.240399e-02
#> [461]  5.640516e-02 -3.692527e-02  6.159552e-02 -1.249692e-01 -1.194905e-01
#> [466]  8.091521e-02 -5.720614e-02 -7.705448e-03  3.410582e-02 -2.241148e-02
#> [471] -4.028940e-02 -3.125432e-02 -9.339451e-03  4.429431e-02 -3.764616e-03
#> [476]  1.199871e-02 -5.461096e-02 -4.797883e-02  5.784693e-02 -1.158380e-01
#> [481] -6.359798e-03  6.380623e-02 -8.815609e-02  7.178401e-02  1.497836e-02
#> [486]  4.881185e-02 -4.038552e-02 -1.597158e-02 -6.431886e-02  1.699684e-02
#> [491]  1.257542e-02 -4.922836e-02 -1.395311e-02  4.467878e-02  4.309284e-03
#> [496]  3.939230e-02  2.805039e-02 -6.143533e-02  3.179899e-02  2.391732e-02
#> [501]  1.968814e-02 -5.768673e-02 -4.759436e-02 -7.710253e-02 -1.130506e-01
#> [506]  1.015805e-01 -4.471082e-02  4.246807e-02  5.352162e-02  7.889673e-02
#> [511] -9.392316e-02  2.939604e-02  2.622415e-02  1.978426e-02 -2.164254e-02
#> [516] -5.807120e-02 -7.402676e-02 -4.182728e-02  5.236821e-02  3.429805e-02
#> [521] -6.230039e-02 -1.279969e-02  5.092644e-02  5.486727e-02 -5.879208e-03
#> [526]  2.699310e-02 -7.883266e-02  6.428682e-02  1.122496e-01  7.192819e-03
#> [531]  2.708921e-02  1.497836e-02  3.160675e-02  5.054197e-02 -2.404548e-02
#> [536]  5.400221e-02  1.344048e-02 -7.796760e-02 -5.614884e-02  1.593954e-02
#> [541] -5.201578e-02  4.073795e-02 -1.116569e-02  8.389486e-02 -5.470708e-02
#> [546]  3.650876e-02  6.870824e-02  3.958453e-02  1.199871e-02 -1.683664e-02
#> [551] -1.433758e-02  3.636459e-03  4.910020e-02 -9.065515e-02 -5.807120e-02
#> [556]  3.285628e-02  7.947344e-02 -1.090137e-01  3.516311e-02  8.634587e-03
#> [561]  3.247181e-02 -5.494737e-03  4.275642e-02  6.294117e-02  3.554758e-02
#> [566] -9.007844e-02 -1.962406e-02  5.150315e-02 -4.259623e-02  8.946970e-02
#> [571] -3.380144e-03  6.745871e-02  3.103005e-02  5.429057e-02  9.456394e-02
#> [576]  5.304103e-02 -3.000479e-02  3.540341e-03  9.033476e-02  8.490410e-04
#> [581] -5.893626e-02 -4.149087e-03  7.769526e-03 -6.648151e-03 -1.540929e-01
#> [586] -6.922087e-02 -3.423397e-02  3.948842e-02  1.488225e-02 -6.297321e-02
#> [591]  8.504828e-02  5.717410e-02  5.438668e-02 -2.913973e-02 -1.558711e-02
#> [596]  5.313715e-02  6.495965e-02  4.371760e-02 -6.076250e-02 -2.250760e-02
#> [601] -2.971644e-02  6.904466e-03  2.771398e-03  1.286377e-02  2.680086e-02
#> [606] -1.307363e-01 -3.572380e-03  6.668977e-02  5.025362e-02  5.563622e-02
#> [611]  1.459389e-02 -6.902864e-02  3.131840e-02 -2.539113e-02 -4.922836e-02
#> [616] -1.231430e-01  4.256419e-02  1.680460e-02  9.211294e-03  6.524800e-02
#> [621]  3.746994e-02  1.642013e-02 -3.221550e-02  9.307412e-03  2.314838e-02
#> [626] -8.854056e-02 -1.142040e-01 -2.279595e-02  1.219095e-02 -8.825221e-02
#> [631]  8.908523e-02  3.622041e-02  4.083407e-02 -3.202326e-02 -8.469585e-02
#> [636] -3.000479e-02  6.784318e-02  4.333313e-02 -4.269234e-02  7.514814e-02
#> [641]  4.563996e-02  3.458641e-02  6.659365e-02 -4.374964e-02 -3.557962e-02
#> [646] -2.039300e-02 -5.461096e-02 -1.923959e-02 -4.096222e-02 -2.846690e-02
#> [651]  6.904466e-03 -9.051097e-03  2.805039e-02 -3.827092e-02  7.288937e-03
#> [656] -3.644468e-02 -1.126661e-01 -3.750198e-02  2.228332e-02 -2.212313e-02
#> [661]  4.131466e-02  7.385055e-03 -1.097345e-02  3.410582e-02 -4.341323e-03
#> [666]  4.083407e-02 -6.979758e-02  3.151063e-02  1.613178e-02  2.141826e-02
#> [671] -2.041703e-01 -3.884763e-02  2.516686e-02  1.247930e-02 -3.990493e-02
#> [676]  2.497462e-02 -1.084370e-01 -4.836330e-02 -6.888446e-04 -1.130506e-01
#> [681]  4.852349e-02  1.488225e-02 -3.433009e-02 -1.222299e-02  9.792807e-02
#> [686]  3.968065e-02 -9.478822e-02  6.712230e-03 -1.938377e-03  2.776204e-02
#> [691] -8.488808e-02 -5.586049e-02  4.121854e-02  1.049447e-01 -2.846690e-02
#> [696]  7.529232e-04 -2.404548e-02  1.449778e-02  4.362148e-02 -1.039675e-02
#> [701]  8.908523e-02 -4.442247e-02 -4.567200e-02  6.519994e-03 -1.056495e-01
#> [706]  5.078226e-03 -7.325782e-02  1.853473e-02  6.505576e-02 -4.663318e-02
#> [711] -3.091791e-03  4.294866e-02 -1.293906e-01  4.842738e-02 -6.071444e-03
#> [716]  8.543275e-02 -6.143533e-02 -5.807120e-02 -8.709879e-02  3.093393e-02
#> [721] -5.590855e-03  1.988038e-02  1.853473e-02  4.064183e-02 -3.019702e-02
#> [726]  6.322953e-02  3.843112e-02  3.372135e-02 -5.489931e-02 -9.190468e-02
#> [731]  5.438668e-02  1.233032e-01  8.264533e-02  3.160675e-02 -3.586798e-02
#> [736] -3.682916e-02 -1.549099e-02 -2.106583e-02  5.352162e-02 -6.816357e-02
#> [741] -7.469959e-02  8.254921e-02  3.295240e-02  3.766218e-02 -1.693276e-02
#> [746]  7.956956e-02  4.621667e-02 -1.616382e-02  7.005389e-02 -3.586798e-02
#> [751]  7.192819e-03 -1.279969e-02 -2.981255e-02  2.189885e-02  4.900408e-02
#> [756] -2.846690e-02  8.552887e-02  3.775830e-02  1.161424e-02  5.592457e-02
#> [761] -4.004910e-04 -5.393813e-02 -5.494737e-03 -6.950922e-02  4.006512e-02
#> [766]  5.073420e-02  2.026485e-02  5.621292e-02  2.093767e-02  6.327758e-03
#> [771]  5.150315e-02 -1.693276e-02  3.141452e-02  4.405401e-03  6.342176e-02
#> [776]  4.765843e-02 -8.777162e-02 -6.662569e-02 -1.090137e-01 -1.895124e-02
#> [781] -1.842259e-03  6.111493e-02  6.217223e-02 -2.856302e-02  6.726648e-02
#> [786]  4.275642e-02 -6.167562e-03 -3.961657e-02 -1.260746e-02  4.073795e-02
#> [791] -2.611202e-03  3.996901e-02 -6.552033e-03  6.178776e-02 -6.864416e-02
#> [796]  2.468627e-02  2.064932e-02 -7.864042e-02 -1.087734e-02  2.535909e-02
#> [801]  6.246058e-02 -1.789394e-02  1.632401e-02  6.399847e-02  4.612055e-02
#> [806]  9.821642e-02 -1.135793e-02  3.545147e-02  2.814651e-02 -1.116569e-02
#> [811]  6.457518e-02  6.880436e-02 -1.799006e-02 -6.825969e-02  1.011000e-01
#> [816]  6.178776e-02  1.466598e-01  2.853098e-02  6.774706e-02  3.348105e-03
#> [821]  6.707424e-02  4.789873e-03  5.236821e-02  3.093393e-02  2.055320e-02
#> [826]  7.293743e-02 -3.067761e-02 -2.452607e-02  8.250115e-03  6.332564e-02
#> [831]  7.149566e-02  3.843112e-02 -8.488808e-02 -2.077748e-02  3.400970e-02
#> [836]  4.785067e-02 -3.836704e-02 -1.885512e-02  3.064557e-02  1.017247e-02
#> [841]  9.917760e-02 -5.614884e-02  9.446783e-02  6.476741e-02  2.218720e-02
#> [846] -7.854430e-02  4.708173e-02  8.351039e-02  4.160301e-02 -1.698563e-01
#> [851]  4.938855e-02 -7.854430e-02  1.228707e-02 -7.897683e-03 -1.799006e-02
#> [856] -2.720295e-01 -3.346503e-02 -2.298819e-02  4.583219e-02 -2.414160e-02
#> [861]  8.730705e-03  4.746620e-02  2.132214e-02  8.860464e-02  5.986540e-02
#> [866]  4.352537e-02  9.119982e-02 -1.032466e-01 -5.980132e-02 -1.042078e-01
#> [871]  5.159927e-02  8.024238e-02  1.016767e-01  4.813902e-02 -2.707319e-03
#> [876] -1.674053e-02  7.226460e-02  6.015375e-02 -1.601964e-05 -6.499169e-02
#> [881]  1.329630e-03 -2.183477e-02  9.571736e-02  1.555507e-02 -2.279595e-02
#> [886] -6.624122e-02  7.985791e-02 -6.566451e-02 -6.335768e-02  4.285254e-02
#> [891]  4.679337e-02  2.290809e-03  1.617984e-03  1.018689e-01 -3.663692e-02
#> [896] -2.193089e-02 -6.556839e-02  9.946595e-02 -1.161264e-01 -4.672929e-02
#> [901] -1.135793e-02 -5.067013e-02  1.632401e-02 -1.702888e-02  2.708921e-02
#> [906]  2.555133e-02  5.294492e-02  5.544398e-02  5.457892e-02  1.082127e-01
#> [911]  7.447531e-02 -1.702888e-02 -7.609330e-03 -3.956851e-03 -4.423023e-02
#> [916]  2.132214e-02  1.094142e-02 -5.980132e-02  1.959202e-02  2.641639e-02
#> [921] -6.744269e-03  6.169164e-02  5.823140e-02  5.746245e-02 -4.821912e-03
#> [926] -6.249262e-02 -2.130612e-03 -5.590855e-03  7.264907e-02  2.949216e-02
#> [931] -3.211938e-02  6.423876e-03 -1.376568e-01  4.294866e-02  1.286377e-02
#> [936]  5.371386e-02  6.611306e-02  1.978426e-02 -4.980507e-02  2.064932e-02
#> [941]  2.161050e-02 -1.231910e-02  4.785067e-02 -2.702514e-02  2.362897e-02
#> [946]  6.476741e-02  2.247556e-02 -2.260371e-02  1.959202e-02  3.420193e-02
#> [951] -9.844070e-02  7.947344e-02  3.468252e-02 -1.097345e-02 -1.981630e-02
#> [956] -2.856302e-02  4.871573e-02  4.727396e-02  4.467878e-02  1.055694e-02
#> [961] -8.133172e-02  5.409833e-02 -8.762744e-03  4.150689e-02 -2.692902e-02
#> [966] -5.624496e-02 -1.361670e-03 -1.365995e-01 -1.030063e-02  3.400970e-02
#> [971] -2.644843e-02 -1.021893e-01 -5.807120e-02  1.670849e-02  7.505202e-02
#> [976]  1.155176e-01  1.411330e-02 -2.606396e-02  5.371386e-02  4.717784e-02
#> [981] -8.037054e-02  1.005233e-01  3.276017e-02  9.235323e-02  7.264907e-02
#> [986] -7.513212e-03  4.304478e-02 -4.201952e-02  9.691883e-03  5.121479e-02
#> [991] -1.053612e-01  2.920381e-02  4.064183e-02  3.795053e-02  7.995403e-02
#> [996] -1.520264e-02  7.774332e-02 -4.557588e-02 -1.049287e-02
#>