Table of correlations and P-values

Table of correlations with associated P-values and df, can be used with regular or independent contrast data

Usage

cor.table(x, cor.method = c("pearson","spearman"),
cor.type=c("standard","contrast"))

Arguments

x

Data frame of data points or contrasts at nodes

cor.method

Correlation method

cor.type

Are data standard or independent contrast values?

References

Garland, T., Jr., P. H. Harvey, and A. R. Ives. 1992. Procedures for the analysis of comparative data using phylogenetically independent contrasts. Systematic Biology 41:18-32. Table of correlations and P-values Table of correlations with associated P-values and df, can be used with regular or independent contrast data

Examples

example_1 <- data.frame(
X = c(4.09434, 3.61092, 2.37024, 2.02815, -1.46968),
Y = c(4.74493, 3.33220, 3.36730, 2.89037, 2.30259))
cor.table(x          = example_1, 
         cor.method = "pearson",
         cor.type   = "contrast")
#> $r
#>           X         Y
#> X 1.0000000 0.8479677
#> Y 0.8479677 1.0000000
#> 
#> $df
#> [1] 4
#> 
#> $P
#>            X          Y
#> X 0.00000000 0.03291371
#> Y 0.03291371 0.00000000
#>