Species' evolutionary distinctiveness

Calculates evolutionary distinctiveness measures for a suite of species by: a) equal splits (Redding and Mooers 2006) b) fair proportions (Isaac et al., 2007). Returns a datafram with species identifiers and species scores.

Usage

evol.distinct(
  tree,
  type = c("equal.splits", "fair.proportion"),
  scale = FALSE,
  use.branch.lengths = TRUE
)

Arguments

tree

an object of class phylo

type

a) equal splits (Redding and Mooers 2006) or b) fair proportions (Isaac et al., 2007)

scale

The scale option refers to whether or not the phylogeny should be scaled to a depth of 1 or, in the case of an ultrametric tree, scaled such that branch lengths are relative.

use.branch.lengths

If use.branch.lengths=FALSE, then all branch lengths are changed to 1.

Note

This function will return a vector of evolutionary distinctivenss for every species in the given tree. If only a subset of values are needed there are two, concetually distinct options: either prune the tree first and then pass the tree in or subset the resulting vector. These two options will provide very different outputs.

References

Redding, D.W. and Mooers, A.O. (2006). Incorporating evolutionary measures into conservation prioritisation. Conservation Biology, 20, 1670-1678.

Isaac, N.J.B., Turvey, S.T., Collen, B., Waterman, C. and Baillie, J.E.M. (2007). Mammals on the EDGE: conservation priorities based on threat and phylogeny. PLoS ONE, 2, e296.

Mark Vellend, William K. Cornwell, Karen Magnuson-Ford, and Arne Mooers. In press. Measuring phylogenetic biodiversity. In: Biological diversity: frontiers in measurement and assessment. Edited by Anne Magurran and Brian McGill.

Author

Karen Magnuson-Ford, Will Cornwell, Arne Mooers, Mark Vellend

Examples

data(phylocom)
evol.distinct(phylocom$phylo, type = 'equal.splits')
#>    Species      w
#> 1      sp1 1.9375
#> 2      sp2 1.9375
#> 3      sp3 1.9375
#> 4      sp4 1.9375
#> 5      sp5 1.9375
#> 6      sp6 1.9375
#> 7      sp7 1.9375
#> 8      sp8 1.9375
#> 9      sp9 1.9375
#> 10    sp10 1.9375
#> 11    sp11 1.9375
#> 12    sp12 1.9375
#> 13    sp13 1.9375
#> 14    sp14 1.9375
#> 15    sp15 1.9375
#> 16    sp16 1.9375
#> 17    sp17 1.9375
#> 18    sp18 1.9375
#> 19    sp19 1.9375
#> 20    sp20 1.9375
#> 21    sp21 1.9375
#> 22    sp22 1.9375
#> 23    sp23 1.9375
#> 24    sp24 1.9375
#> 25    sp25 1.9375
#> 26    sp26 1.9375
#> 27    sp27 1.9375
#> 28    sp28 1.9375
#> 29    sp29 1.9375
#> 30    sp30 1.9375
#> 31    sp31 1.9375
#> 32    sp32 1.9375