Phylogenetic index of beta-diversity PhyloSor

Fraction of branch-length shared between two communities

phylosor(samp, tree)

Arguments

samp

Community data matrix

tree

Object of class phylo - a rooted phylogeny

Value

A distance object of the PhyloSor index of similarity between communities, the fraction of PD (branch-length) shared between two samples

Note

The root of the supplied tree is included in calculations of PhyloSor. The supplied tree must be rooted. Single-species samples will be assigned a PD value equal to the distance from the root to the present.

Warning

The phylosor of all samples will include the branch length connecting taxa in those samples and the root of the supplied tree. The root of the supplied tree may not be spanned by any taxa in the sample. If you want the root of your tree to correspond to the most recent ancestor of the taxa actually present in your sample, you should prune the tree before running phylosor:

prunedTree <- prune.sample(sample,tree)

References

Bryant, J.B., Lamanna, C., Morlon, H., Kerkhoff, A.J., Enquist, B.J., Green, J.L. 2008. Microbes on mountainsides: Contrasting elevational patterns of bacterial and plant diversity. Proceedings of the National Academy of Sciences 105 Supplement 1: 11505-11511

See also

Author

Helene Morlon morlon.helene@gmail.com and Steven Kembel steve.kembel@gmail.com

Examples

data(phylocom)
phylosor(phylocom$sample, phylocom$phylo)
#>            clump1   clump2a   clump2b    clump4      even
#> clump2a 0.5454545                                        
#> clump2b 0.5294118 0.5142857                              
#> clump4  0.3157895 0.5641026 0.6000000                    
#> even    0.3478261 0.3829787 0.4166667 0.6923077          
#> random  0.3720930 0.4090909 0.4444444 0.4897959 0.6315789