Standardized effect size of MNTD

Standardized effect size of mean nearest taxon distances in communities. When used with a phylogenetic distance matrix, equivalent to -1 times the Nearest Taxon Index (NTI).

Usage

ses.mntd(
  samp,
  dis,
  null.model = c("taxa.labels", "richness", "frequency", "sample.pool", "phylogeny.pool",
    "independentswap", "trialswap"),
  abundance.weighted = FALSE,
  runs = 999,
  iterations = 1000
)

Arguments

samp

Community data matrix

dis

Distance matrix (generally a phylogenetic distance matrix)

null.model

Null model to use (see Details section for description)

abundance.weighted

Should mean nearest taxon distances for each species be weighted by species abundance? (default = FALSE)

runs

Number of randomizations

iterations

Number of iterations to use for each randomization (for independent swap and trial null models)

Value

A data frame of results for each community

ntaxa

Number of taxa in community

mntd.obs

Observed MNTD in community

mntd.rand.mean

Mean MNTD in null communities

mntd.rand.sd

Standard deviation of MNTD in null communities

mntd.obs.rank

Rank of observed MNTD vs. null communities

mntd.obs.z

Standardized effect size of MNTD vs. null communities (= (mntd.obs - mntd.rand.mean) / mntd.rand.sd, equivalent to -NTI)

mntd.obs.p

P-value (quantile) of observed MNTD vs. null communities (= mntd.obs.rank / runs + 1)

runs

Number of randomizations

Details

The metric used by this function has also been referred to as MNND (mean nearest neighbour distance), and the function was named ses.mnnd in picante versions < 0.7.

Currently implemented null models (arguments to null.model):

taxa.labels

Shuffle distance matrix labels (across all taxa included in distance matrix)

richness

Randomize community data matrix abundances within samples (maintains sample species richness)

frequency

Randomize community data matrix abundances within species (maintains species occurence frequency)

sample.pool

Randomize community data matrix by drawing species from pool of species occurring in at least one community (sample pool) with equal probability

phylogeny.pool

Randomize community data matrix by drawing species from pool of species occurring in the distance matrix (phylogeny pool) with equal probability

independentswap

Randomize community data matrix with the independent swap algorithm (Gotelli 2000) maintaining species occurrence frequency and sample species richness

trialswap

Randomize community data matrix with the trial-swap algorithm (Miklos & Podani 2004) maintaining species occurrence frequency and sample species richness

References

Webb, C.O., Ackerly, D.D., and Kembel, S.W. 2008. Phylocom: software for the analysis of phylogenetic community structure and trait evolution. Version 4.0.1. http://www.phylodiversity.net/phylocom/.

See also

Author

Steven Kembel steve.kembel@gmail.com

Examples

data(phylocom)
ses.mntd(phylocom$sample, cophenetic(phylocom$phylo),null.model="taxa.labels")
#>         ntaxa mntd.obs mntd.rand.mean mntd.rand.sd mntd.obs.rank mntd.obs.z
#> clump1      8        2       4.707207    0.6196782           1.0 -4.3687305
#> clump2a     8        2       4.730480    0.6318963           1.0 -4.3210894
#> clump2b     8        2       4.727728    0.6558790           1.0 -4.1588885
#> clump4      8        2       4.736236    0.6454312           1.0 -4.2393927
#> even        8        6       4.734484    0.6480544         999.0  1.9527922
#> random      8        5       4.703954    0.6632395         631.5  0.4463637
#>         mntd.obs.p runs
#> clump1      0.0010  999
#> clump2a     0.0010  999
#> clump2b     0.0010  999
#> clump4      0.0010  999
#> even        0.9990  999
#> random      0.6315  999