Standardized effect size of MPD

Standardized effect size of mean pairwise distances in communities. When used with a phylogenetic distance matrix, equivalent to -1 times the Nearest Relative Index (NRI).

ses.mpd(
  samp,
  dis,
  null.model = c("taxa.labels", "richness", "frequency", "sample.pool", "phylogeny.pool",
    "independentswap", "trialswap"),
  abundance.weighted = FALSE,
  runs = 999,
  iterations = 1000
)

Arguments

samp

Community data matrix

dis

Distance matrix (generally a phylogenetic distance matrix)

null.model

Null model to use (see Details section for description)

abundance.weighted

Should mean nearest taxon distances for each species be weighted by species abundance? (default = FALSE)

runs

Number of randomizations

iterations

Number of iterations to use for each randomization (for independent swap and trial null models)

Value

A data frame of results for each community

ntaxa

Number of taxa in community

mpd.obs

Observed mpd in community

mpd.rand.mean

Mean mpd in null communities

mpd.rand.sd

Standard deviation of mpd in null communities

mpd.obs.rank

Rank of observed mpd vs. null communities

mpd.obs.z

Standardized effect size of mpd vs. null communities (= (mpd.obs - mpd.rand.mean) / mpd.rand.sd, equivalent to -NRI)

mpd.obs.p

P-value (quantile) of observed mpd vs. null communities (= mpd.obs.rank / runs + 1)

runs

Number of randomizations

Details

Currently implemented null models (arguments to null.model):

taxa.labels

Shuffle distance matrix labels (across all taxa included in distance matrix)

richness

Randomize community data matrix abundances within samples (maintains sample species richness)

frequency

Randomize community data matrix abundances within species (maintains species occurence frequency)

sample.pool

Randomize community data matrix by drawing species from pool of species occurring in at least one community (sample pool) with equal probability

phylogeny.pool

Randomize community data matrix by drawing species from pool of species occurring in the distance matrix (phylogeny pool) with equal probability

independentswap

Randomize community data matrix with the independent swap algorithm (Gotelli 2000) maintaining species occurrence frequency and sample species richness

trialswap

Randomize community data matrix with the trial-swap algorithm (Miklos & Podani 2004) maintaining species occurrence frequency and sample species richness

References

Webb, C.O., Ackerly, D.D., and Kembel, S.W. 2008. Phylocom: software for the analysis of phylogenetic community structure and trait evolution. Version 4.0.1. http://www.phylodiversity.net/phylocom/.

See also

Author

Steven Kembel steve.kembel@gmail.com

Examples

data(phylocom)
ses.mpd(phylocom$sample, cophenetic(phylocom$phylo),null.model="taxa.labels")
#>         ntaxa  mpd.obs mpd.rand.mean mpd.rand.sd mpd.obs.rank   mpd.obs.z
#> clump1      8 4.857143      8.318390   0.3234931          1.0 -10.6996011
#> clump2a     8 6.000000      8.323323   0.3095141          1.0  -7.5063569
#> clump2b     8 7.142857      8.334120   0.3303909          6.5  -3.6056155
#> clump4      8 8.285714      8.344702   0.3069320        342.5  -0.1921844
#> even        8 8.857143      8.339411   0.2943029        997.0   1.7591806
#> random      8 8.428571      8.322537   0.3255376        548.0   0.3257215
#>         mpd.obs.p runs
#> clump1     0.0010  999
#> clump2a    0.0010  999
#> clump2b    0.0065  999
#> clump4     0.3425  999
#> even       0.9970  999
#> random     0.5480  999