Standardized effect size of PD

Standardized effect size of phylogenetic diversity (Faith's PD) in communities.

ses.pd(
  samp,
  tree,
  null.model = c("taxa.labels", "richness", "frequency", "sample.pool", "phylogeny.pool",
    "independentswap", "trialswap"),
  runs = 999,
  iterations = 1000,
  include.root = TRUE
)

Arguments

samp

Community data matrix

tree

Phylogeny (phylo object)

null.model

Null model to use (see Details section for description)

runs

Number of randomizations

iterations

Number of iterations to use for each randomization (for independent swap and trial null models)

include.root

Include distance to root node in calculation of PD (see documentation in pd function

Value

A data frame of results for each community

ntaxa

Number of taxa in community

pd.obs

Observed PD in community

pd.rand.mean

Mean PD in null communities

pd.rand.sd

Standard deviation of PD in null communities

pd.obs.rank

Rank of observed PD vs. null communities

pd.obs.z

Standardized effect size of PD vs. null communities (= (pd.obs - pd.rand.mean) / pd.rand.sd)

pd.obs.p

P-value (quantile) of observed PD vs. null communities (= mpd.obs.rank / runs + 1)

runs

Number of randomizations

Details

Currently implemented null models (arguments to null.model):

taxa.labels

Shuffle taxa labels across tips of phylogeny (across all taxa included in phylogeny)

richness

Randomize community data matrix abundances within samples (maintains sample species richness)

frequency

Randomize community data matrix abundances within species (maintains species occurence frequency)

sample.pool

Randomize community data matrix by drawing species from pool of species occurring in at least one community (sample pool) with equal probability

phylogeny.pool

Randomize community data matrix by drawing species from pool of species occurring in the phylogeny (phylogeny pool) with equal probability

independentswap

Randomize community data matrix with the independent swap algorithm (Gotelli 2000) maintaining species occurrence frequency and sample species richness

trialswap

Randomize community data matrix with the trial-swap algorithm (Miklos & Podani 2004) maintaining species occurrence frequency and sample species richness

References

Webb, C.O., Ackerly, D.D., and Kembel, S.W. 2008. Phylocom: software for the analysis of phylogenetic community structure and trait evolution. Version 4.0.1. http://www.phylodiversity.net/phylocom/.

Proches, S., Wilson, J.R.U. and Cowling, R.M. 2006. How much evolutionary history in a 10 x 10m plot? Proceedings of Royal Society of London B, Biological Sciences 273:1143-1148.

See also

Author

Steven Kembel steve.kembel@gmail.com

Examples

data(phylocom)
ses.pd(phylocom$sample, phylocom$phylo, null.model="taxa.labels", runs=99)
#>         ntaxa pd.obs pd.rand.mean pd.rand.sd pd.obs.rank   pd.obs.z pd.obs.p
#> clump1      8     16     26.20202   1.647384         1.0 -6.1928617    0.010
#> clump2a     8     17     26.33333   1.584362         1.0 -5.8909083    0.010
#> clump2b     8     18     26.68687   1.542900         1.0 -5.6302218    0.010
#> clump4      8     22     26.38384   1.454810         2.0 -3.0133418    0.020
#> even        8     30     26.40404   1.603053        99.5  2.2431942    0.995
#> random      8     27     26.54545   1.364811        60.5  0.3330465    0.605
#>         runs
#> clump1    99
#> clump2a   99
#> clump2b   99
#> clump4    99
#> even      99
#> random    99