Species co-occurrence distances

Compute interspecific distances based on patterns of species co-occurrence in communities.

species.dist(x, metric = c("cij", "jaccard", "checkerboard", "doij"))

Arguments

x

Community data matrix

metric

Co-occurrence metric to use (see Details section for description)

Value

A dist object with co-occurrences among all species pairs

Details

Currently implemented co-occurrence measures (arguments to metric):

cij

Schoener's index of co-occurrence

jaccard

Jaccard index of co-occurrence

checkerboard

Checkerboard index of co-occurrence

doij

DOij index of co-occurrence

References

Hardy, O.J. 2008. Testing the spatial phylogenetic structure of local communities: statistical performances of different null models and test statistics on a locally neutral community. Journal of Ecology 96:914-926.

See also

Author

Steven Kembel steve.kembel@gmail.com