Taxonomic distinctiveness sensu Vane-Wright or May

Taxic diversity: Vane-Wright et al., 1991 and May 1990 which accounts for polytomies by counting the number of branches descending from each node that lies on the path from a spp tip to the root (not just counting the number of nodes).

Usage

tax.distinctiveness(tree, type = c("Vane-Wright", "May"))

Arguments

tree

an object of class phylo

type

specify "Vane-Wright" or "May"

References

Vane-Wright, R.I., Humphries, C.J. and Williams, P.H. (1991). What to protect? - Systematics and the agony of choice. Biological Conservation, 55, 235-254.

May, R.M. (1990). Taxonomy as destiny. Nature, 347, 129-130.

Mark Vellend, William K. Cornwell, Karen Magnuson-Ford, and Arne Mooers. In press. Measuring phylogenetic biodiversity In: Biological diversity: frontiers in measurement and assessment. Edited by Anne Magurran and Brian McGill.

Author

Karen Magnuson-Ford, Will Cornwell, Arne Mooers, Mark Vellend

Examples

data(phylocom)
tax.distinctiveness(phylocom$phylo,type = 'Vane-Wright')
#>    Species       w
#> 1      sp1 0.03125
#> 2      sp2 0.03125
#> 3      sp3 0.03125
#> 4      sp4 0.03125
#> 5      sp5 0.03125
#> 6      sp6 0.03125
#> 7      sp7 0.03125
#> 8      sp8 0.03125
#> 9      sp9 0.03125
#> 10    sp10 0.03125
#> 11    sp11 0.03125
#> 12    sp12 0.03125
#> 13    sp13 0.03125
#> 14    sp14 0.03125
#> 15    sp15 0.03125
#> 16    sp16 0.03125
#> 17    sp17 0.03125
#> 18    sp18 0.03125
#> 19    sp19 0.03125
#> 20    sp20 0.03125
#> 21    sp21 0.03125
#> 22    sp22 0.03125
#> 23    sp23 0.03125
#> 24    sp24 0.03125
#> 25    sp25 0.03125
#> 26    sp26 0.03125
#> 27    sp27 0.03125
#> 28    sp28 0.03125
#> 29    sp29 0.03125
#> 30    sp30 0.03125
#> 31    sp31 0.03125
#> 32    sp32 0.03125