Unweighted UniFrac distance between communities

Calculates unweighted UniFrac, a phylogenetic beta diversity metric of the the unique (non-shared) fraction of total phylogenetic diversity (branch-length) between two communities.

unifrac(comm, tree)

Arguments

comm

Community data matrix

tree

Object of class phylo - a rooted phylogeny

Value

A dist object of the unweighted UniFrac distances between communities (the unique (non-shared) fraction of total phylogenetic diversity (branch-length) between two communities).

Note

The supplied tree must be rooted. Single-species samples will be assigned a PD value equal to the distance from the root to the present.

Warning

The UniFrac distance between samples will include the branch length connecting taxa in those samples and the root of the supplied tree. The root of the supplied tree may not be spanned by any taxa in the sample. If you want the root of your tree to correspond to the most recent ancestor of the taxa actually present in your samples, you should prune the tree before running unifrac: prunedTree <- prune.sample(sample,tree)

References

Lozupone, C., Hamady, M., and Knight, R. 2006. UniFrac - an online tool for comparing microbial community diversity in a phylogenetic context. BMC Bioinformatics 7:371.

See also

Author

Steven Kembel steve.kembel@gmail.com

Examples

data(phylocom)
unifrac(phylocom$sample, phylocom$phylo)
#>            clump1   clump2a   clump2b    clump4      even
#> clump2a 0.6250000                                        
#> clump2b 0.6400000 0.6538462                              
#> clump4  0.8125000 0.6071429 0.5714286                    
#> even    0.7894737 0.7631579 0.7368421 0.4705882          
#> random  0.7714286 0.7428571 0.7142857 0.6756757 0.5384615