Calculates inter-community mean nearest taxon distance

Calculates MNTD (mean nearest taxon distance) separating taxa in two communities, a measure of phylogenetic beta diversity

Usage

comdistnt(comm, dis, abundance.weighted = FALSE, exclude.conspecifics = FALSE)

Arguments

comm

Community data matrix

dis

Interspecific distance matrix

abundance.weighted

Should mean nearest taxon distances from each species to species in the other community be weighted by species abundance? (default = FALSE)

exclude.conspecifics

Should conspecific taxa in different communities be exclude from MNTD calculations? (default = FALSE)

Value

Distance object of MNTD values separating each pair of communities.

Details

This metric has also been referred to as MNND (mean nearest neighbour distance).

This function calculates a measure of phylogenetic beta diversity: the average phylogenetic distance to the most similar taxon or individual in the other community for taxa or individuals in two communities.

References

C.O. Webb, D.D. Ackerly, and S.W. Kembel. 2008. Phylocom: software for the analysis of phylogenetic community structure and trait evolution. Bioinformatics 18:2098-2100.

See also

Author

Steven Kembel steve.kembel@gmail.com

Examples

data(phylocom)
comdistnt(phylocom$sample, cophenetic(phylocom$phylo), abundance.weighted=FALSE)
#> [1] "Dropping taxa from the distance matrix because\n     they are not present in the community data:"
#> [1] "sp16" "sp23" "sp27" "sp28" "sp30" "sp31" "sp32"
#>         clump1 clump2a clump2b clump4  even
#> clump2a  3.500                             
#> clump2b  4.000   4.500                     
#> clump4   5.500   3.500   3.000             
#> even     4.750   4.500   4.000  2.000      
#> random   4.625   4.250   3.875  3.875 2.750