Correlations between species co-occurrence and phylogenetic distances

Calculates measures of community phylogenetic structure (correlation between co-occurrence and phylogenetic distance) to patterns expected under various null models

comm.phylo.cor(
  samp,
  phylo,
  metric = c("cij", "checkerboard", "jaccard", "doij"),
  null.model = c("sample.taxa.labels", "pool.taxa.labels", "frequency", "richness",
    "independentswap", "trialswap"),
  runs = 999,
  ...
)

Arguments

samp

Community data matrix

phylo

Phylogenetic tree

metric

Metric of co-occurrence to use (see species.dist)

null.model

Null model to use (see Details section for description)

runs

Number of runs (randomizations)

...

Additional arguments to randomizeMatrix

Value

A list with elements:

obs.corr

Observed co-occurrence/phylogenetic distance correlation

obs.corr.p

P-value of observed correlation (standard P-value for correlation coefficient, not based on comparison with randomizations)

obs.rank

Rank of observed correlation vs. random

runs

Number of runs (randomizations)

obs.rand.p

P-value of observed correlation vs. randomizations (= obs.rank / (runs + 1))

random.corrs

A vector of random correlation calculated for each run

Details

Currently implemented null models (arguments to null.model):

sample.taxa.labels

Shuffle phylogeny tip labels (only within set of taxa present in community data)

pool.taxa.labels

Shuffle phylogeny tip labels (across all taxa included in phylogenetic tree)

frequency

Randomize community data matrix abundances within species (maintains species occurence frequency)

richness

Randomize community data matrix abundances within samples (maintains sample species richness)

independentswap

Randomize community data matrix maintaining species occurrence frequency and site richnessing using independent swap

trialswap

Randomize community data matrix maintaining species occurrence frequency and site richnessing using trial swap

References

Cavender-Bares J., D.A. Ackerly, D. Baum and F.A. Bazzaz. 2004. Phylogenetic overdispersion in Floridian oak communities, American Naturalist, 163(6):823-843.

See also

Author

Steven Kembel steve.kembel@gmail.com

Examples

data(phylocom)
comm.phylo.cor(phylocom$sample, phylocom$phylo, metric="cij",null.model="sample.taxa.labels")
#> $obs.corr
#> [1] -0.390639
#> 
#> $obs.corr.p
#> [1] 2.240087e-12
#> 
#> $obs.rank
#> [1] 1
#> 
#> $runs
#> [1] 999
#> 
#> $obs.rand.p
#> [1] 0.001
#> 
#> $random.corrs
#>   [1]  8.072297e-02 -1.938377e-03 -5.624496e-02  6.620918e-02 -1.731723e-02
#>   [6] -4.629676e-03  9.691883e-03 -7.993801e-03  1.007635e-02 -8.123560e-02
#>  [11] -4.855553e-02  1.141720e-01  9.100758e-02  5.063809e-02  5.611680e-02
#>  [16]  3.410582e-02  9.980237e-03 -2.683290e-02  5.544398e-02  3.881559e-02
#>  [21]  1.151812e-02  7.577290e-03 -3.529127e-02 -9.723922e-03 -5.288084e-02
#>  [26] -5.783091e-03 -6.393439e-02 -3.860734e-03  8.101133e-02 -1.606770e-02
#>  [31] -1.597158e-02  2.103379e-02 -5.884014e-02 -5.932073e-02 -4.605647e-02
#>  [36]  2.507074e-02  3.756606e-02  3.987289e-02 -4.336517e-02  6.294117e-02
#>  [41]  1.228707e-02  3.170287e-02 -1.183852e-02  5.400221e-02 -2.154642e-02
#>  [46]  4.727396e-02  4.775455e-02  4.501519e-03  5.034973e-02 -4.845942e-02
#>  [51]  1.680460e-02  1.565119e-02 -1.068510e-02 -4.341323e-03  8.927746e-02
#>  [56] -1.048806e-01 -1.650023e-03 -4.615259e-02 -4.269234e-02 -1.601964e-05
#>  [61]  2.064932e-02  2.756980e-02  7.293743e-02 -2.048912e-02  4.458266e-02
#>  [66] -5.826344e-02 -8.690656e-02  1.255139e-01 -3.490680e-02 -1.246809e-01
#>  [71]  6.601694e-02 -7.729477e-02 -1.645217e-02 -5.095848e-02 -1.058898e-02
#>  [76]  5.006138e-02 -4.990118e-02 -1.654829e-02  1.046083e-02 -3.144656e-02
#>  [81] -5.432260e-02 -8.373467e-02 -7.787148e-02  3.910395e-02  7.096701e-03
#>  [86]  1.536284e-02  2.036096e-02  5.390609e-02  9.273771e-02 -2.442995e-02
#>  [91] -1.481817e-02  3.468252e-02 -2.567949e-02  2.170661e-02  1.420942e-02
#>  [96] -5.566825e-02  4.016124e-02 -7.652583e-02  1.334436e-02  4.054571e-02
#> [101]  4.660114e-02 -6.931699e-02  3.093393e-02 -1.904735e-02  4.737008e-02
#> [106]  4.150689e-02 -2.058524e-02  6.428682e-02 -6.931699e-02  1.906337e-03
#> [111] -1.399636e-01 -4.009716e-02  9.677465e-02 -1.204517e-01  2.257167e-02
#> [116] -5.682167e-02 -3.702139e-02 -8.277349e-02  8.966193e-02  3.410582e-02
#> [121]  2.098573e-03  2.680086e-02 -4.437441e-03  2.305226e-02 -7.393065e-02
#> [126]  5.159927e-02 -2.241148e-02 -1.212206e-01 -1.443370e-02  4.717784e-02
#> [131] -3.673304e-02  2.545521e-02  6.419070e-02 -6.383827e-02  6.582471e-02
#> [136] -9.771981e-04 -1.767767e-01 -6.071444e-03 -3.971269e-02 -1.434239e-01
#> [141] -3.404174e-02  7.418696e-02  4.117048e-03 -1.991241e-02  4.948467e-02
#> [146] -1.000747e-01  1.920755e-02  7.063060e-02  2.507074e-02  2.641639e-02
#> [151]  1.863084e-02  2.324450e-02 -1.102632e-01  5.169538e-02 -6.167562e-03
#> [156]  5.621292e-02  4.352537e-02  6.841989e-02 -3.279221e-02  2.122603e-02
#> [161] -2.760184e-02  7.880062e-02 -1.289581e-02 -1.404923e-02 -4.211564e-02
#> [166]  8.264533e-02  5.227209e-02 -5.874403e-02  3.189511e-02 -5.518767e-02
#> [171] -2.202701e-02  4.189136e-02  6.630530e-02  2.151438e-02  8.033850e-02
#> [176] -7.224858e-03  4.583219e-02  5.246433e-02  1.058097e-01  5.938481e-02
#> [181]  1.382495e-02  6.231641e-03  5.832752e-02 -2.981255e-02  1.565119e-02
#> [186] -2.418966e-03  5.361774e-02 -9.017456e-02  1.101350e-01 -8.661820e-02
#> [191] -4.629676e-03  4.117048e-03  4.323701e-02  6.904466e-03 -1.702888e-02
#> [196]  4.938855e-02  2.805039e-02 -6.508780e-02  3.497088e-02 -8.181231e-02
#> [201]  1.122496e-01  2.002455e-03 -3.096597e-02 -6.230039e-02 -3.029314e-02
#> [206] -4.629676e-03 -1.938377e-03 -6.249262e-02  3.141452e-02 -6.652957e-02
#> [211]  3.074169e-02  5.847169e-03 -8.474390e-03  1.632401e-02  5.174344e-03
#> [216]  3.732576e-03  1.401719e-02  6.265282e-02  4.266031e-02  5.366580e-03
#> [221] -3.721363e-02 -9.627805e-03 -2.354086e-01 -5.634108e-02 -5.807120e-02
#> [226] -3.423397e-02 -2.279595e-02 -2.418966e-03 -1.425588e-01  3.814277e-02
#> [231] -3.259997e-02 -1.741335e-02 -1.169434e-03 -4.865165e-02  6.572859e-02
#> [236] -4.509529e-02 -1.625994e-02  2.353285e-02 -3.668498e-03 -9.844070e-02
#> [241] -2.644843e-02  4.650502e-02 -9.036680e-02  6.423876e-03  4.333313e-02
#> [246] -1.575532e-01  5.871199e-02 -4.278846e-02 -6.167562e-03 -9.449987e-02
#> [251]  4.083407e-02  6.178776e-02 -6.566451e-02 -8.796385e-02  1.295989e-02
#> [256] -2.529501e-02 -1.347252e-02 -2.952420e-02  6.712230e-03 -8.421526e-02
#> [261] -3.135044e-02 -5.249637e-02  3.708547e-02  3.179899e-02 -4.086611e-02
#> [266]  2.622415e-02 -1.277566e-01  6.784318e-02 -4.596035e-02  7.312966e-02
#> [271] -3.038926e-02 -2.375713e-02  2.386927e-03 -8.954980e-03 -5.855179e-02
#> [276]  6.231641e-03  2.016873e-02 -1.299193e-02 -1.856676e-02  2.314838e-02
#> [281]  4.198748e-02  1.344048e-02 -1.818229e-02  5.342550e-02  4.958079e-02
#> [286] -5.067013e-02 -1.635605e-02 -1.911944e-01 -4.711377e-02 -7.354617e-02
#> [291] -1.153574e-01 -3.284027e-03  3.996901e-02 -1.027660e-01  4.919632e-02
#> [296] -2.366101e-02 -8.325408e-02 -1.625994e-02  1.151812e-02 -6.998981e-02
#> [301]  8.043462e-02 -7.609330e-03 -5.590855e-03  6.044211e-02  2.641639e-02
#> [306]  1.052330e-01 -4.725794e-03  3.679712e-02 -1.320819e-01  5.746245e-02
#> [311]  1.018689e-01 -4.086611e-02 -3.548351e-02  5.505951e-02 -4.932448e-02
#> [316] -9.574940e-02 -5.903238e-02  7.466755e-02 -8.186037e-03  1.968814e-02
#> [321] -6.037803e-02 -8.623373e-02  4.910020e-02 -7.844818e-02 -6.701016e-02
#> [326]  1.411330e-02 -3.413786e-02  9.821642e-02  7.288937e-03  6.265282e-02
#> [331]  9.110370e-02  1.834249e-02  4.756232e-02  6.073046e-02 -3.798257e-02
#> [336]  6.880436e-02 -3.394562e-02  7.928121e-02  3.170287e-02 -4.394188e-02
#> [341]  4.266031e-02  3.227958e-02  4.256419e-02  2.612804e-02 -3.173491e-02
#> [346] -6.816357e-02  6.717036e-02 -9.757564e-02 -1.655310e-01  3.449029e-02
#> [351] -4.701765e-02 -2.721737e-02 -2.856302e-02 -1.538046e-01 -1.039675e-02
#> [356] -4.836330e-02 -2.394936e-02 -1.241522e-02 -6.071444e-03  2.122603e-02
#> [361] -2.356489e-02  8.937358e-02  9.884119e-03 -9.339451e-03 -7.993801e-03
#> [366]  2.564745e-02 -4.115446e-02 -2.039300e-02 -5.413037e-02 -4.028940e-02
#> [371] -1.111283e-01 -6.585675e-02  3.670100e-02  5.294492e-02 -3.187909e-03
#> [376]  1.738131e-02 -5.384202e-02  5.025362e-02  6.909271e-02  3.160675e-02
#> [381] -5.605273e-02 -9.498045e-02  9.302606e-02 -4.355741e-02  4.256419e-02
#> [386] -1.712500e-02 -1.168953e-01 -4.153893e-02  1.747743e-02  3.900783e-02
#> [391] -9.632610e-02 -7.325782e-02 -1.023815e-01  7.091895e-02 -3.010091e-02
#> [396]  1.997649e-02  2.007261e-02  5.659739e-02 -3.000479e-02  1.469001e-02
#> [401] -1.305441e-01  4.515937e-02 -1.395311e-02 -6.345380e-02 -4.557588e-02
#> [406] -5.711002e-02  4.789873e-03  5.851975e-02 -5.240025e-02  3.828694e-03
#> [411] -4.490306e-02  2.093767e-02 -1.347732e-01  1.065306e-02 -8.238902e-02
#> [416]  6.524800e-02  3.189511e-02 -5.259248e-02  1.161424e-02 -2.904361e-02
#> [421]  3.977677e-02  3.122228e-02 -3.654080e-02  2.516686e-02  1.526672e-02
#> [426] -5.307307e-02  1.238318e-02  9.629407e-02 -2.577560e-02 -1.943183e-02
#> [431]  4.515937e-02  2.228332e-02 -2.515084e-03  4.419819e-02  5.400221e-02
#> [436]  4.198748e-02  8.072297e-02  1.555507e-02 -2.154642e-02  2.362897e-02
#> [441] -1.549099e-02  2.708921e-02  1.488225e-02  1.267154e-02 -5.105460e-02
#> [446] -6.172368e-02  9.398724e-02 -3.971269e-02 -1.039194e-01  2.814651e-02
#> [451]  7.514814e-02 -2.769796e-02  3.352911e-02 -2.009984e-01 -7.114323e-02
#> [456]  1.670849e-02  2.872322e-02  1.219095e-02  2.795427e-02  2.785816e-02
#> [461]  6.784318e-02 -3.043732e-04  7.005389e-02 -9.171245e-02  1.272440e-01
#> [466]  9.523677e-02  1.305601e-02 -3.682916e-02 -4.519141e-02 -5.393813e-02
#> [471] -4.557588e-02 -1.616382e-02 -7.421900e-02 -2.327654e-02  4.016124e-02
#> [476]  5.794304e-02 -5.951297e-02 -5.470708e-02  7.918509e-02  3.227958e-02
#> [481] -8.873279e-02  2.372509e-02  8.850852e-02  2.170661e-02 -4.821912e-03
#> [486]  1.805414e-02 -6.936505e-03 -1.683664e-02 -6.162756e-02 -3.567574e-02
#> [491]  2.098573e-03  4.198748e-02 -9.325033e-02  9.812030e-02  2.867516e-03
#> [496]  3.429805e-02 -2.981255e-02 -3.144656e-02  3.314464e-02 -2.962032e-02
#> [501] -6.595286e-02  1.738131e-02 -1.203075e-02  7.178401e-02 -2.995673e-03
#> [506]  4.804291e-02  8.841240e-02 -9.786399e-02  4.833126e-02  8.543275e-02
#> [511] -2.061888e-01  7.409084e-02  2.555133e-02 -4.749824e-02 -2.212313e-02
#> [516]  1.882308e-02  1.171036e-02 -2.442995e-02 -5.855179e-02  5.150315e-02
#> [521] -8.652208e-02  6.515188e-02  8.524051e-02 -8.700267e-02  9.499647e-03
#> [526]  8.850852e-02  1.089816e-01 -1.295829e-01  2.353285e-02  1.853473e-02
#> [531] -7.758312e-02  4.535161e-02  3.035722e-02  5.284880e-02 -1.594755e-01
#> [536] -1.875900e-02  4.285254e-02 -6.167562e-03 -2.942808e-02  3.237570e-02
#> [541] -3.284027e-03 -2.366101e-02  1.882308e-02  7.524426e-02 -1.049287e-02
#> [546] -5.720614e-02 -1.914347e-02  4.405401e-03  2.353285e-02  5.900034e-02
#> [551]  1.055694e-02  2.747369e-02  7.707050e-02 -1.972018e-02 -1.650023e-03
#> [556]  2.747369e-02  1.690072e-02  5.880810e-02  1.795802e-02  3.564370e-02
#> [561] -1.125700e-01 -3.663692e-02 -1.491429e-02 -2.952420e-02  3.881559e-02
#> [566]  6.322953e-02  4.631278e-02 -4.096222e-02 -8.671432e-02 -1.435200e-01
#> [571]  3.251987e-03 -7.224858e-03  6.476741e-02 -5.057401e-02  6.765095e-02
#> [576]  1.026859e-02  7.341802e-02 -2.164254e-02 -9.757564e-02 -3.096597e-02
#> [581] -6.960534e-02 -1.779782e-02  3.708547e-02 -3.721363e-02  7.111119e-02
#> [586] -1.203075e-02 -9.051097e-03 -6.744269e-03  4.597637e-03  7.053448e-02
#> [591]  6.327758e-03  1.228707e-02 -4.336517e-02  2.161050e-02 -5.739838e-02
#> [596] -1.376087e-02 -2.923585e-02  2.103379e-02  1.449778e-02  1.119613e-01
#> [601]  3.006887e-02  3.295240e-02 -1.866288e-02  4.645696e-04 -5.509155e-02
#> [606] -6.744269e-03 -9.276974e-02  3.756606e-02  4.525549e-02 -1.424627e-01
#> [611] -5.432260e-02  4.885991e-03  1.002349e-01 -2.899555e-03  4.544772e-02
#> [616]  6.044211e-02  3.756606e-02 -6.124309e-02 -7.897683e-03  4.405401e-03
#> [621] -9.627805e-03 -8.642597e-02  1.411330e-02  4.217972e-02  8.826823e-03
#> [626] -2.981255e-02 -3.298444e-02  1.036471e-02 -4.499917e-02 -4.115446e-02
#> [631] -1.501040e-02 -3.077373e-02 -4.192340e-02 -1.054573e-01  4.419819e-02
#> [636] -4.788271e-02 -8.671432e-02 -1.470763e-01  9.595765e-03  3.064557e-02
#> [641]  3.324076e-02 -4.245205e-03  7.197625e-02 -4.451858e-02 -8.796385e-02
#> [646]  7.005389e-02 -2.789020e-02 -2.356489e-02  3.708547e-02  8.346233e-03
#> [651]  7.769526e-03  5.919258e-02  6.313341e-02  5.496339e-02 -1.203075e-02
#> [656] -7.671806e-02  8.153998e-03 -3.029314e-02  2.516686e-02 -3.461845e-02
#> [661]  9.033476e-02  2.728145e-02 -1.414534e-02  3.924812e-03  5.943287e-03
#> [666]  4.583219e-02  2.382121e-02 -2.750572e-02 -8.613761e-02 -2.279595e-02
#> [671] -1.837453e-02  9.639018e-02  2.843486e-02  2.074544e-02  1.074918e-02
#> [676] -8.719491e-02  4.496713e-02  7.649379e-02  3.045334e-02 -8.392690e-02
#> [681] -4.634482e-02  3.122228e-02 -1.414534e-02  4.804291e-02 -3.077373e-02
#> [686]  8.697063e-02  2.708921e-02 -3.807869e-02  3.775830e-02 -7.897683e-03
#> [691] -2.798631e-02  9.581348e-02 -1.272761e-01 -6.701016e-02 -1.501040e-02
#> [696] -6.455916e-03  4.141077e-02 -1.078122e-02  3.641265e-02  5.217597e-02
#> [701] -4.663318e-02 -2.481442e-02  8.053074e-02  6.986166e-02  3.506700e-02
#> [706]  1.718908e-02  1.315213e-02  3.324076e-02  5.275268e-02 -3.577186e-02
#> [711]  9.879313e-02  6.486353e-02 -2.500666e-02  5.179150e-02 -1.021893e-01
#> [716] -7.815983e-02  1.103753e-02  6.246058e-02 -4.437441e-03  6.226835e-02
#> [721] -1.328509e-01  4.602443e-02 -3.259997e-02  7.101507e-02  2.478239e-02
#> [726]  8.346233e-03  9.403530e-03 -5.864791e-02 -4.288458e-02 -8.604149e-02
#> [731] -8.248513e-02 -1.019971e-01 -4.605647e-02  7.716661e-02 -8.858862e-03
#> [736]  3.304852e-02 -4.461470e-02 -1.760559e-02  9.494842e-02 -1.058898e-02
#> [741]  1.028301e-01 -2.202701e-02 -2.760184e-02  7.534037e-02 -1.362150e-01
#> [746]  8.033850e-02  2.574356e-02  8.072297e-02 -1.866288e-02 -4.894001e-02
#> [751]  6.063434e-02  3.929618e-02 -6.624122e-02  1.363271e-02  1.060020e-01
#> [756]  3.477864e-02 -4.245205e-03 -4.149087e-03  3.795053e-02  2.420568e-02
#> [761] -3.807869e-02 -9.911352e-02  9.360277e-02  8.014627e-02 -1.183852e-02
#> [766]  6.428682e-02 -8.940562e-02  3.891171e-02 -8.834832e-02  8.057880e-03
#> [771]  1.109040e-01 -2.087840e-01 -4.624871e-02  8.129968e-02  8.158803e-02
#> [776]  1.863084e-02  7.783944e-02 -3.961657e-02  7.673408e-03  3.314464e-02
#> [781]  5.943287e-03  2.929992e-02 -4.605647e-02 -7.662195e-02 -1.608212e-01
#> [786] -1.115127e-01  4.861961e-02 -7.691030e-02  3.573982e-02 -7.633359e-02
#> [791]  4.919632e-02  1.267154e-02 -6.970146e-02 -7.633359e-02 -1.368878e-01
#> [796]  1.632401e-02  5.751051e-03 -4.374964e-02 -5.634108e-02 -4.557588e-02
#> [801]  7.192819e-03  4.169913e-02 -2.032092e-01  3.276017e-02  1.181128e-01
#> [806]  6.082658e-02  1.363271e-02 -1.546696e-01 -2.356489e-02 -1.164147e-01
#> [811]  3.756606e-02 -2.068136e-02 -2.452607e-02  2.218720e-02 -1.039675e-02
#> [816] -6.979758e-02 -1.529395e-01  2.161050e-02  9.595765e-03 -5.018954e-02
#> [821]  9.307412e-03 -1.683184e-01 -1.429433e-01 -5.461096e-02 -2.193089e-02
#> [826]  6.808348e-03 -2.375713e-02 -3.961657e-02 -1.601964e-05 -4.432635e-02
#> [831] -1.760559e-02 -1.062262e-01  4.496713e-02  1.190259e-02  5.342550e-02
#> [836]  6.870824e-02  1.526672e-02 -3.086985e-02  3.622041e-02  1.046083e-02
#> [841] -1.328028e-02  1.094142e-02 -6.840387e-03 -8.373467e-02  8.178027e-02
#> [846] -3.971269e-02 -1.356864e-02 -9.147215e-03  2.910769e-02  4.698561e-02
#> [851] -3.356115e-02  3.641265e-02 -2.885137e-02  4.842738e-02  3.343299e-02
#> [856]  1.834249e-02  5.717410e-02 -6.701016e-02  3.766218e-02 -3.000479e-02
#> [861]  5.140703e-02  1.949590e-02  3.160675e-02  7.322578e-02 -1.271799e-01
#> [866]  4.592831e-02 -5.855179e-02  8.793181e-02 -3.394562e-02  1.329630e-03
#> [871]  8.543275e-02  6.423876e-03  9.788001e-03 -1.145404e-02 -1.827841e-02
#> [876]  3.151063e-02 -1.991241e-02  4.881185e-02  4.698561e-02 -8.666626e-03
#> [881]  5.131091e-02  2.632027e-02 -4.692153e-02  3.083781e-02  5.698187e-02
#> [886]  4.890797e-02  3.602817e-02  3.208734e-02 -1.549099e-02 -3.572380e-03
#> [891]  7.139954e-02 -4.653706e-02  7.245684e-02  9.019058e-03 -4.576812e-02
#> [896]  4.650502e-02  6.841989e-02 -4.336517e-02  3.199122e-02 -5.134295e-02
#> [901] -1.501040e-02 -3.135044e-02  2.785816e-02 -5.413037e-02  1.363271e-02
#> [906]  5.227209e-02 -1.103593e-01 -5.720614e-02 -4.259623e-02  1.037913e-01
#> [911]  8.697063e-02 -4.720988e-02  8.057880e-03  2.545521e-02 -1.856676e-02
#> [916]  4.405401e-03  6.178776e-02 -3.250385e-02  8.043462e-02 -9.147215e-03
#> [921] -4.413411e-02 -1.126181e-02 -3.606021e-02 -3.644468e-02  2.401344e-02
#> [926]  1.065787e-01  4.314089e-02  3.525923e-02 -1.553905e-03  8.178027e-02
#> [931]  4.573608e-02 -8.296572e-02  5.304103e-02 -4.576812e-02  2.814651e-02
#> [936]  6.332564e-02  1.062903e-01  5.775081e-02  2.766592e-02 -4.932448e-02
#> [941]  7.673408e-03 -5.398619e-03 -7.421900e-02 -2.385325e-02  4.405401e-03
#> [946] -4.624871e-02  2.901157e-02  1.382495e-02 -1.250653e-01  2.372509e-02
#> [951] -3.956851e-03 -3.452233e-02 -9.435569e-03 -4.341323e-03  1.088855e-01
#> [956]  5.928869e-02  3.564370e-02  4.083407e-02  2.295615e-02  9.466006e-02
#> [961] -1.074758e-01 -8.325408e-02  7.803167e-02  5.784693e-02  5.640516e-02
#> [966]  3.059752e-03 -2.034494e-03  9.225712e-02  2.766592e-02 -7.056652e-02
#> [971] -7.224858e-03 -2.260371e-02  7.399472e-02 -1.808618e-02  1.699684e-02
#> [976] -1.041117e-01 -2.606396e-02  8.408710e-02  8.456769e-02 -2.433384e-02
#> [981]  6.169164e-02 -7.104711e-02 -3.346503e-02  4.333313e-02  1.411330e-02
#> [986]  5.602069e-02 -3.336891e-02  3.444223e-03 -4.725794e-03 -3.702139e-02
#> [991]  8.850852e-02 -5.783091e-03  9.168041e-02  8.389486e-02 -1.616382e-02
#> [996]  3.583594e-02 -2.356489e-02  6.611306e-02 -2.025363e-01
#>