Quantile regression slopes between species co-occurrence and phylogenetic distances

Calculates measures of community phylogenetic structure (quantile regression between co-occurrence and phylogenetic distance) to patterns expected under various null models

Usage

comm.phylo.qr(
  samp,
  phylo,
  metric = c("cij", "checkerboard", "jaccard", "doij"),
  null.model = c("sample.taxa.labels", "pool.taxa.labels", "frequency", "richness",
    "independentswap", "trialswap"),
  quant = 0.75,
  runs = 999,
  show.plot = FALSE,
  ...
)

Arguments

samp

Community data matrix

phylo

Phylogenetic tree

metric

Metric of co-occurrence to use (see species.dist)

null.model

Null model to use (see Details section for description)

quant

Quantile of slope to be fit (using rq)

runs

Number of runs (randomizations)

show.plot

Option to display a plot of co-occurrence versus phylogenetic distance with quantile regression slope fit

...

Additional arguments to randomizeMatrix

Value

A list with elements:

obs.qr.intercept

Observed co-occurrence/phylogenetic distance quantile regression intercept

obs.qr.slope

Observed co-occurrence/phylogenetic distance quantile regression slope

obs.qr.slope.p

P-value of observed quantile regression slope significance versus null model (calculated based on comparison with randomizations)

obs.rank

Rank of observed quantile regression slope vs. random

runs

Number of runs (randomizations)

random.qr.slopes

A vector of quantile regression slopes calculated for each randomization

Details

This function fits a quantile regression of co-occurrence versus phylogenetic distances separating species, and compares observed patterns to the patterns expected under some null model. The quantile regressions are fit using the rq function from the quantreg package.

Currently implemented null models (arguments to null.model):

sample.taxa.labels

Shuffle phylogeny tip labels (only within set of taxa present in community data)

pool.taxa.labels

Shuffle phylogeny tip labels (across all taxa included in phylogenetic tree)

frequency

Randomize community data matrix abundances within species (maintains species occurence frequency)

richness

Randomize community data matrix abundances within samples (maintains sample species richness)

independentswap

Randomize community data matrix maintaining species occurrence frequency and site richnessing using independent swap

trialswap

Randomize community data matrix maintaining species occurrence frequency and site richnessing using trial swap

References

Cavender-Bares J., D.A. Ackerly, D. Baum and F.A. Bazzaz. 2004. Phylogenetic overdispersion in Floridian oak communities, American Naturalist, 163(6):823-843. Slingsby, J. A. and G. A. Verboom. 2006. Phylogenetic relatedness limits coexistence at fine spatial scales: evidence from the schoenoid sedges (Cyperaceae: Schoeneae) of the Cape Floristic Region, South Africa. The American Naturalist 168:14-27.

See also

Author

Steven Kembel steve.kembel@gmail.com

Examples

data(phylocom)
comm.phylo.qr(phylocom$sample, phylocom$phylo, metric="cij",
  null.model="sample.taxa.labels", runs=99)
#> Loading required namespace: quantreg
#> $obs.qr.intercept
#> [1] 0.8666667
#> 
#> $obs.qr.slope
#> [1] -0.06666667
#> 
#> $obs.qr.slope.p
#> [1] 0.01
#> 
#> $obs.rank
#> [1] 1
#> 
#> $runs
#> [1] 99
#> 
#> $random.qr.slopes
#>  [1]  1.387779e-17  1.387779e-17 -1.387779e-17 -4.857226e-17  2.081668e-17
#>  [6]  2.081668e-17 -1.850372e-17 -3.122502e-17 -4.163336e-17 -1.387779e-17
#> [11]  1.445603e-17  1.757853e-17  2.081668e-17 -6.938894e-17  1.041667e-02
#> [16] -2.775558e-17  2.081668e-17 -6.938894e-18  1.387779e-17 -3.469447e-17
#> [21]  1.387779e-17 -1.387779e-17  3.989864e-17  1.387779e-17  2.081668e-17
#> [26]  3.561966e-17  1.387779e-17  2.081668e-17  1.041667e-02 -1.850372e-17
#> [31] -1.850372e-17 -5.204170e-17 -1.619075e-17  2.081668e-17  1.387779e-17
#> [36]  1.388889e-02 -1.666667e-02 -1.850372e-17  1.387779e-17  1.387779e-17
#> [41]  2.081668e-17  2.081668e-17 -1.619075e-17  1.214306e-17 -5.551115e-17
#> [46] -1.619075e-17 -1.041667e-02 -4.857226e-17  6.245005e-17  1.041667e-02
#> [51]  1.387779e-17 -6.938894e-18 -1.111111e-02  3.816392e-17 -1.619075e-17
#> [56]  1.387779e-17 -1.387779e-17  6.245005e-17  2.081668e-17  6.245005e-17
#> [61]  2.081668e-17  2.081668e-17  1.387779e-17  1.041667e-02  2.081668e-17
#> [66] -4.857226e-17 -1.619075e-17  6.245005e-17  1.561251e-17  3.561966e-17
#> [71]  6.245005e-17  3.561966e-17  1.387779e-17 -2.777778e-02  1.474515e-17
#> [76]  6.245005e-17  1.041667e-02 -5.551115e-17 -2.775558e-17  2.081668e-17
#> [81]  3.989864e-17 -1.850372e-17  1.388889e-02 -6.245005e-17  1.434038e-17
#> [86]  2.081668e-17  1.041667e-02  6.245005e-17 -1.387779e-17  1.387779e-17
#> [91] -1.850372e-17  2.081668e-17  2.081668e-17 -1.387779e-17  2.081668e-17
#> [96] -2.255141e-17 -1.387779e-17 -6.938894e-17 -6.765422e-17
#> 
#> $quantile
#> [1] 0.75
#>