Convert community data matrix to Phylocom sample

Converts a community data matrix to a Phylocom database-format community sample

Usage

matrix2sample(z)

Arguments

z

Community data matrix

Value

Phylocom database-format community sample

References

Webb, C.O., Ackerly, D.D., and Kembel, S.W. 2008. Phylocom: software for the analysis of phylogenetic community structure and trait evolution. Version 4.0.1. http://www.phylodiversity.net/phylocom/.

Author

Steven Kembel steve.kembel@gmail.com and Cam Webb cwebb@oeb.harvard.edu

Examples

data(phylocom)
matrix2sample(phylocom$sample)
#>       plot abund   id
#> 1   clump1     1  sp1
#> 2   clump1     1  sp2
#> 3   clump1     1  sp3
#> 4   clump1     1  sp4
#> 5   clump1     1  sp5
#> 6   clump1     1  sp6
#> 7   clump1     1  sp7
#> 8   clump1     1  sp8
#> 9  clump2a     1  sp1
#> 10 clump2a     2 sp10
#> 11 clump2a     2 sp11
#> 12 clump2a     2 sp12
#> 13 clump2a     1  sp2
#> 14 clump2a     1  sp3
#> 15 clump2a     1  sp4
#> 16 clump2a     2  sp9
#> 17 clump2b     1  sp1
#> 18 clump2b     2 sp17
#> 19 clump2b     2 sp18
#> 20 clump2b     2 sp19
#> 21 clump2b     1  sp2
#> 22 clump2b     2 sp20
#> 23 clump2b     1  sp3
#> 24 clump2b     1  sp4
#> 25  clump4     1  sp1
#> 26  clump4     1 sp10
#> 27  clump4     2 sp17
#> 28  clump4     2 sp18
#> 29  clump4     1  sp2
#> 30  clump4     2 sp25
#> 31  clump4     2 sp26
#> 32  clump4     1  sp9
#> 33    even     1  sp1
#> 34    even     1 sp13
#> 35    even     1 sp17
#> 36    even     1 sp21
#> 37    even     1 sp25
#> 38    even     1 sp29
#> 39    even     1  sp5
#> 40    even     1  sp9
#> 41  random     1 sp12
#> 42  random     4 sp14
#> 43  random     2 sp15
#> 44  random     3 sp17
#> 45  random     1  sp2
#> 46  random     1 sp22
#> 47  random     2 sp24
#> 48  random     2  sp5