Phylogenetic Community Dissimilarity

Pairwise dissimilarity in phylogenetic community composition that is partitioned into a nonphylogenetic and a phylogenetic component.

Usage

pcd(comm, tree, PSVmncd = NULL, PSVpool = NULL, reps = 10^4)

Arguments

comm

Community data matrix

tree

Object of class phylo or a phylogenetic covariance matrix

PSVmncd

Vector of null mean conditional phylogenetic species variability (PSV) values

PSVpool

The standard, unconditional PSV calculated for the species pool

reps

The number of random draws from the species pool used to produce PSVmncd

Value

The function returns a list with items:

PCD

A square matrix of PCD values

PCDc

A square matrix of PCDc values

PCDp

A square matrix of PCDp values

PSVmncd

A vector of null mean conditional PSV values used to calculate PCD

PSVpool

The unconditional PSV of the species pool used to calculate PCD

Details

Phylogenetic community dissimilarity (PCD) is the pairwise differences between communities derived by asking how much of the variance among species in the values of a hypothetical nonselected trait in one community can be predicted by the known trait values of species in another community. PCD is partitioned into a nonphylogenetic component that reflects shared species between communities (PCDc) and a phylogenetic component that reflects the evolutionary relationships among nonshared species (PCDp). In order to compare communities that vary in species richness, the metric is standardized under the assumption that the species in communities are selected at random from the species pool. The analyses here define the species pool as the list of all species in the set of communities in comm, but the species pool can be defined under any hypothesis of community assembly either by manipulating the code or inputting a user defined PSVmncd and PSVpool.

Note

The sampling procedure used to standardize PCD and produce PSVmncd and PSVpool can be slow.

References

Ives A.R. & Helmus M.R. (2010). Phylogenetic metrics of community similarity. The American Naturalist, 176, E128-E142.

See also

Author

Anthony Ives arives@wisc.edu and Matthew Helmus mrhelmus@gmail.com

Examples

# \donttest{
data(phylocom)
pcd(phylocom$sample, phylocom$phylo)
#> $PCD
#>            clump1   clump2a   clump2b    clump4      even
#> clump2a 1.2083222                                        
#> clump2b 1.1339414 1.1077813                              
#> clump4  1.6195854 0.9916052 0.9303750                    
#> even    1.5100144 1.3699766 1.2736179 0.6979766          
#> random  1.5079310 1.3565054 1.2404003 1.2004128 0.9254718
#> 
#> $PCDc
#>            clump1   clump2a   clump2b    clump4      even
#> clump2a 0.7352941                                        
#> clump2b 0.7352941 0.7352941                              
#> clump4  1.1029412 0.7352941 0.7352941                    
#> even    1.1029412 1.1029412 1.1029412 0.7352941          
#> random  1.1029412 1.1029412 1.1029412 1.1029412 1.1029412
#> 
#> $PCDp
#>            clump1   clump2a   clump2b    clump4      even
#> clump2a 1.6433182                                        
#> clump2b 1.5421604 1.5065826                              
#> clump4  1.4684241 1.3485831 1.2653099                    
#> even    1.3690797 1.2421121 1.1547469 0.9492481          
#> random  1.3671908 1.2298983 1.1246296 1.0883743 0.8390944
#> 
#> $PSVmncd
#> [1] 0.7509380 0.6812656 0.6277342 0.5732205 0.5200471 0.4734049 0.4344402
#> [8] 0.3953108
#> 
#> $PSVpool
#> [1] 0.8233333
#> 
# }

pcd(phylocom$sample,phylocom$phylo)
#> $PCD
#>            clump1   clump2a   clump2b    clump4      even
#> clump2a 1.2085084                                        
#> clump2b 1.1341162 1.1079520                              
#> clump4  1.6198350 0.9917580 0.9305183                    
#> even    1.5102471 1.3701877 1.2738142 0.6980841          
#> random  1.5081634 1.3567145 1.2405914 1.2005978 0.9256144
#> 
#> $PCDc
#>            clump1   clump2a   clump2b    clump4      even
#> clump2a 0.7352941                                        
#> clump2b 0.7352941 0.7352941                              
#> clump4  1.1029412 0.7352941 0.7352941                    
#> even    1.1029412 1.1029412 1.1029412 0.7352941          
#> random  1.1029412 1.1029412 1.1029412 1.1029412 1.1029412
#> 
#> $PCDp
#>            clump1   clump2a   clump2b    clump4      even
#> clump2a 1.6435714                                        
#> clump2b 1.5423980 1.5068147                              
#> clump4  1.4686504 1.3487909 1.2655049                    
#> even    1.3692907 1.2423035 1.1549248 0.9493944          
#> random  1.3674015 1.2300878 1.1248029 1.0885420 0.8392237
#> 
#> $PSVmncd
#> [1] 0.7491920 0.6794767 0.6234253 0.5669527 0.5178110 0.4732155 0.4304833
#> [8] 0.3952499
#> 
#> $PSVpool
#> [1] 0.8233333
#>