Null PhyloSor values of phylogenetic beta-diversity

PhyloSor values obtained by randomization for different choices of null models

Usage

phylosor.rnd(
  samp,
  tree,
  cstSor = TRUE,
  null.model = c("taxa.labels", "frequency", "richness", "independentswap", "trialswap"),
  runs = 999,
  iterations = 1000
)

Arguments

samp

Community data matrix

tree

Object of class phylo - a rooted phylogeny

cstSor

TRUE if the Sorensen similarity should be kept constant across communities. FALSE otherwise

null.model

Null model to use (see Details section)

runs

Number of randomizations

iterations

Number of iterations to use for each randomization (for independent swap and trial null models)

Value

A list of length the number of runs. Each element of the list is a distance matrix containing the PhyloSor values of phylogenetic beta-diversity obtained by randomization

Details

Currently implemented null models (arguments to null.model):

taxa.labels

Shuffle community data matrix labels. Maintains species richness in each community and species shared between communities. Should be used with cstSor=TRUE

frequency

Randomize community data matrix abundances within species (maintains species occurence frequency). Does not maintain species richness in communities nor species shared between communities. Can only be used with cstSor=FALSE

richness

With cstSor=TRUE: For each pair of community, maintains species richness in each community and species shared between communities. Sample in the species pool with equal probability; With cstSor=FALSE: Maintains species richness in each community, does not maintain species shared between communities. Sample in the species pool with equal probability

independentswap

Randomize community data matrix with the independent swap algorithm (Gotelli 2000) maintaining species occurrence frequency and sample species richness. Can only be used with cstSor=FALSE

trialswap

Randomize community data matrix with the trial-swap algorithm (Miklos & Podani 2004) maintaining species occurrence frequency and sample species richness. Can only be used with cstSor=FALSE

References

Bryant, J.B., Lamanna, C., Morlon, H., Kerkhoff, A.J., Enquist, B.J., Green, J.L. 2008. Microbes on mountainsides: Contrasting elevational patterns of bacterial and plant diversity. Proceedings of the National Academy of Sciences 105 Supplement 1: 11505-11511

See also

Author

Helene Morlon morlon.helene@gmail.com and Steven Kembel steve.kembel@gmail.com

Examples

data(phylocom)
phylosor.rnd(phylocom$sample,phylocom$phylo,cstSor=TRUE,null.model="richness",runs=5)
#> [[1]]
#>            clump1   clump2a   clump2b    clump4      even
#> clump2a 0.6923077                                        
#> clump2b 0.7407407 0.6153846                              
#> clump4  0.5490196 0.7600000 0.7450980                    
#> even    0.5490196 0.5185185 0.5714286 0.7058824          
#> random  0.5200000 0.5818182 0.5490196 0.5660377 0.4400000
#> 
#> [[2]]
#>            clump1   clump2a   clump2b    clump4      even
#> clump2a 0.7924528                                        
#> clump2b 0.6666667 0.6000000                              
#> clump4  0.5098039 0.6538462 0.5600000                    
#> even    0.5200000 0.6181818 0.7017544 0.6808511          
#> random  0.5882353 0.5000000 0.6785714 0.5555556 0.4897959
#> 
#> [[3]]
#>            clump1   clump2a   clump2b    clump4      even
#> clump2a 0.7272727                                        
#> clump2b 0.5882353 0.7307692                              
#> clump4  0.5384615 0.6938776 0.6666667                    
#> even    0.5714286 0.6071429 0.4583333 0.6909091          
#> random  0.5660377 0.5555556 0.6181818 0.5098039 0.5769231
#> 
#> [[4]]
#>            clump1   clump2a   clump2b    clump4      even
#> clump2a 0.6938776                                        
#> clump2b 0.6415094 0.6792453                              
#> clump4  0.6530612 0.6415094 0.7450980                    
#> even    0.6181818 0.3829787 0.5818182 0.6538462          
#> random  0.5090909 0.4489796 0.4905660 0.5357143 0.6296296
#> 
#> [[5]]
#>            clump1   clump2a   clump2b    clump4      even
#> clump2a 0.7200000                                        
#> clump2b 0.7017544 0.7636364                              
#> clump4  0.5555556 0.7058824 0.7169811                    
#> even    0.5614035 0.6071429 0.5098039 0.6274510          
#> random  0.5490196 0.5882353 0.6415094 0.5200000 0.4705882
#>