Permutations to Test for Phylogenetic Signal in Community Composition

Randomize sample/community data matrices to create null distributions of given metrics

Usage

phylostruct(
  samp,
  tree,
  env = NULL,
  metric = c("psv", "psr", "pse", "psc", "sppregs"),
  null.model = c("frequency", "richness", "independentswap", "trialswap"),
  runs = 100,
  it = 1000,
  alpha = 0.05,
  fam = "binomial"
)

Arguments

samp

community data matrix, species as columns, communities as rows

tree

phylo tree object or a phylogenetic covariance matrix

env

environmental data matrix

metric

if metric="psv", "psr", "pse", or "psc" compares the observed mean of the respective metric to a null distribution at a given alpha; if metric="sppregs" compares the three correlations produced by sppregs to null distributions

null.model

permutation procedure used to create the null distribution, see randomizeMatrix

runs

the number of permutations to create the distribution, a rule of thumb is (number of communities)/alpha

it

the number of swaps for the independent and trial-swap null models, see randomizeMatrix

alpha

probability value to compare the observed mean/correlations to a null distribution

fam

as in sppregs

Value

metric

metric used

null.model

permutation used

runs

number of permutations

it

number of swaps if applicable

obs

observed mean value of a particular metric or the three observed correlations from sppregs

mean.null

mean(s) of the null distribution(s)

quantiles.null

quantiles of the null distribution(s) to compare to obs; determined by alpha

phylo.structure

if obs less than (alpha/2), phylo.structure="underdispersed"; if obs greater than (1-alpha/2), phylo.structure="overdispersed"; otherwise phylo.structure="random" and NULL if metric="sppregs"

nulls

null values of the distribution(s)

Details

The function creates null distributions for the psd set of metrics and for the correlations of sppregs from observed community data sets.

References

Helmus M.R., Bland T.J., Williams C.K. & Ives A.R. (2007a) Phylogenetic measures of biodiversity. American Naturalist, 169, E68-E83

Helmus M.R., Savage K., Diebel M.W., Maxted J.T. & Ives A.R. (2007b) Separating the determinants of phylogenetic community structure. Ecology Letters, 10, 917-925

Gotelli N.J. (2000) Null model analysis of species co-occurrence patterns. Ecology, 81, 2606-2621

See also

Author

Matthew Helmus mrhelmus@gmail.com

Examples

data(phylocom)
phylostruct(phylocom$sample,phylocom$phylo, metric = 'psv')
#> $metric
#> [1] "psv"
#> 
#> $null.model
#> [1] "frequency"
#> 
#> $runs
#> [1] 100
#> 
#> $it
#> [1] NA
#> 
#> $mean.obs
#> [1] 0.7261905
#> 
#> $mean.null
#> [1] 0.7890694
#> 
#> $quantiles.null
#>      2.5%     97.5% 
#> 0.7601567 0.8102884 
#> 
#> $phylo.structure
#> [1] "underdispersed"
#> 
#> $null.means
#>   [1] 0.8063300 0.7792761 0.7828114 0.7640720 0.7941799 0.8020370 0.8011472
#>   [8] 0.8022559 0.7965608 0.7859585 0.8075493 0.7790909 0.7883842 0.7968335
#>  [15] 0.7912338 0.7866955 0.7811640 0.7834127 0.7894180 0.8058201 0.7867725
#>  [22] 0.8044974 0.8008995 0.7826190 0.7878307 0.7913757 0.7798822 0.7650794
#>  [29] 0.7870370 0.7997660 0.7852381 0.7724074 0.7815344 0.7661111 0.7813781
#>  [36] 0.7903030 0.7864550 0.8025878 0.7925782 0.8014286 0.7558323 0.8096825
#>  [43] 0.7664526 0.7793338 0.8051587 0.7955291 0.8156902 0.8107407 0.7895575
#>  [50] 0.7941126 0.7822222 0.7805339 0.8040741 0.7834055 0.7914021 0.7702930
#>  [57] 0.8032323 0.8005026 0.8026190 0.7737037 0.7936340 0.7860558 0.7791799
#>  [64] 0.7871429 0.8230736 0.7646296 0.7817508 0.7665242 0.7918519 0.7683333
#>  [71] 0.7568254 0.7977778 0.8054473 0.7957672 0.7905820 0.8068783 0.7909259
#>  [78] 0.7655988 0.7615873 0.7588624 0.7971429 0.8048052 0.7632251 0.7961183
#>  [85] 0.7779101 0.7991895 0.7794541 0.8036772 0.8038552 0.7950938 0.7938119
#>  [92] 0.7928083 0.8097884 0.7930736 0.7884031 0.7914021 0.7851948 0.7759307
#>  [99] 0.8033165 0.7962169
#>