Phylogenetic Species Diversity Metrics

Calculate the bounded phylogenetic biodiversity metrics: phylogenetic species variability, richness, evenness and clustering for one or multiple samples.

psd(samp, tree, compute.var = TRUE, scale.vcv = TRUE)

Arguments

samp

Community data matrix

tree

A phylo tree object or a phylogenetic covariance matrix

compute.var

Computes the expected variances for PSV and PSR for each community

scale.vcv

Scale the phylogenetic covariance matrix to bound the metric between 0 and 1

Value

Returns a dataframe of the respective phylogenetic species diversity metric values

Details

Phylogenetic species variability (PSV) quantifies how phylogenetic relatedness decreases the variance of a hypothetical unselected/neutral trait shared by all species in a community. The expected value of PSV is statistically independent of species richness, is one when all species in a sample are unrelated (i.e., a star phylogeny) and approaches zero as species become more related. PSV is directly related to mean phylogenetic distance, except except calculated on a scaled phylogenetic covariance matrix. The expected variance around PSV for any sample of a particular species richness can be approximated. To address how individual species contribute to the mean PSV of a data set, the function psv.spp gives signed proportions of the total deviation from the mean PSV that occurs when all species are removed from the data set one at a time. The absolute values of these “species effects” tend to positively correlate with species prevalence.

Phylogenetic species richness (PSR) is the number of species in a sample multiplied by PSV. It can be considered the species richness of a sample after discounting by species relatedness. The value is maximum at the species richness of the sample, and decreases towards zero as relatedness increases. The expected variance around PSR for any sample of a particular species richness can be approximated.

Phylogenetic species evenness (PSE) is the metric PSV modified to incorporate relative species abundances. The maximum attainable value of PSE (i.e., 1) occurs only if species abundances are equal and species phylogeny is a star. PSE essentially grafts each individual of a species onto the tip of the phylogeny of its species with branch lengths of zero.

Phylogenetic species clustering (PSC) is a metric of the branch tip clustering of species across a sample's phylogeny. As PSC increases to 1, species are less related to one another the tips of the phylogeny. PSC is directly related to mean nearest neighbor distance.

Note

These metrics are bounded either between zero and one (PSV, PSE, PSC) or zero and species richness (PSR); but the metrics asymptotically approach zero as relatedness increases. Zero can be assigned to communities with less than two species, but conclusions drawn from assigning communities zero values need be carefully explored for any data set. The data sets need not be species-community data sets but may be any sample data set with an associated phylogeny.

References

Helmus M.R., Bland T.J., Williams C.K. & Ives A.R. (2007) Phylogenetic measures of biodiversity. American Naturalist, 169, E68-E83

See also

Author

Matthew Helmus mrhelmus@gmail.com

Examples

data(phylocom)
psd(phylocom$sample, phylocom$phylo)
#>               PSV     var.PSV PSC      PSR    var.PSR       PSE SR
#> clump1  0.4857143 0.001055303 0.2 3.885714 0.06753941 0.4857143  8
#> clump2a 0.6000000 0.001055303 0.2 4.800000 0.06753941 0.5650794  8
#> clump2b 0.7142857 0.001055303 0.2 5.714286 0.06753941 0.6666667  8
#> clump4  0.8285714 0.001055303 0.2 6.628571 0.06753941 0.7936508  8
#> even    0.8857143 0.001055303 0.6 7.085714 0.06753941 0.8857143  8
#> random  0.8428571 0.001055303 0.5 6.742857 0.06753941 0.8125000  8