Species co-occurrence distances

Compute interspecific distances based on patterns of species co-occurrence in communities.

Usage

species.dist(x, metric = c("cij", "jaccard", "checkerboard", "doij"))

Arguments

x

Community data matrix

metric

Co-occurrence metric to use (see Details section for description)

Value

A dist object with co-occurrences among all species pairs

Details

Currently implemented co-occurrence measures (arguments to metric):

cij

Schoener's index of co-occurrence

jaccard

Jaccard index of co-occurrence

checkerboard

Checkerboard index of co-occurrence

doij

DOij index of co-occurrence

References

Hardy, O.J. 2008. Testing the spatial phylogenetic structure of local communities: statistical performances of different null models and test statistics on a locally neutral community. Journal of Ecology 96:914-926.

See also

Author

Steven Kembel steve.kembel@gmail.com

Examples

data(varespec)
species.dist(varespec,metric = 'cij')
#>               Callvulg      Empenigr      Rhodtome      Vaccmyrt      Vaccviti
#> Empenigr  1.756502e-01                                                        
#> Rhodtome  1.132911e-01  2.575702e-01                                          
#> Vaccmyrt  1.601638e-01  3.513117e-01  5.779025e-01                            
#> Vaccviti  2.027309e-01  7.663255e-01  2.283193e-01  2.908455e-01              
#> Pinusylv  2.155312e-01  4.441467e-01  1.056669e-01  1.685996e-01  5.597841e-01
#> Descflex  1.138764e-01  2.677447e-01  5.587264e-01  4.003338e-01  2.559189e-01
#> Betupube  6.896552e-02  1.845516e-01  5.457236e-01  3.647652e-01  1.745264e-01
#> Vacculig  1.202316e-01  2.862488e-01  2.181340e-01  2.060338e-01  2.630721e-01
#> Diphcomp  3.447381e-02  1.890502e-01  8.343266e-03  9.268389e-03  2.260554e-01
#> Dicrsp    4.277735e-02  1.240721e-01  1.481481e-02  4.690199e-02  1.533849e-01
#> Dicrfusc  3.528209e-01  3.098824e-01  2.318467e-01  2.194478e-01  3.149446e-01
#> Dicrpoly  9.071770e-02  2.593654e-01  5.097613e-01  3.782874e-01  3.383095e-01
#> Hylosple  4.437542e-04  1.538161e-01  3.301549e-01  4.770262e-01  1.364757e-01
#> Pleuschr  2.444202e-01  4.825559e-01  2.787815e-01  4.849751e-01  5.349324e-01
#> Polypili  8.675394e-02  1.850111e-01  0.000000e+00  3.968888e-02  2.776043e-01
#> Polyjuni  8.916655e-02  3.319005e-01  2.527076e-02  8.784673e-02  3.517378e-01
#> Polycomm  2.816901e-02  3.372683e-01  5.352113e-01  4.618418e-01  2.441552e-01
#> Pohlnuta  2.356987e-01  6.311544e-01  1.762822e-01  2.548024e-01  7.087972e-01
#> Ptilcili  9.456588e-02  2.973737e-01  4.950942e-01  3.428982e-01  3.340061e-01
#> Barbhatc  2.194357e-02  1.553885e-01  5.101797e-01  3.314860e-01  1.477874e-01
#> Cladarbu  3.591381e-01  4.382275e-01  4.123878e-02  1.038197e-01  4.862856e-01
#> Cladrang  2.630157e-01  4.704138e-01  5.156315e-02  1.097960e-01  4.924273e-01
#> Cladstel  1.630677e-01  3.834420e-01  3.645035e-02  8.339877e-02  3.951291e-01
#> Cladunci  2.561461e-01  2.900756e-01  5.862756e-02  8.655644e-02  4.239232e-01
#> Cladcocc  3.832891e-01  5.178024e-01  7.644362e-02  1.640886e-01  5.938362e-01
#> Cladcorn  2.309287e-01  5.906786e-01  1.028939e-01  1.608348e-01  6.676097e-01
#> Cladgrac  2.693907e-01  6.382227e-01  1.289658e-01  2.157880e-01  7.651177e-01
#> Cladfimb  2.818864e-01  7.068798e-01  1.724847e-01  2.495399e-01  7.495300e-01
#> Cladcris  3.118448e-01  4.636471e-01  7.527767e-02  1.235043e-01  5.399289e-01
#> Cladchlo  5.067979e-02  4.372304e-01  2.411019e-01  2.851305e-01  4.702913e-01
#> Cladbotr  1.515987e-01  3.496633e-01  5.693708e-01  4.277976e-01  3.874131e-01
#> Cladamau  1.286887e-02  9.809856e-02  0.000000e+00  2.220446e-16  9.082646e-02
#> Cladsp    3.062800e-01  3.033018e-01  8.526634e-02  1.246530e-01  3.472362e-01
#> Cetreric  3.001504e-01  3.332133e-01  1.389882e-02  6.570148e-02  4.019052e-01
#> Cetrisla  1.644197e-01  3.748241e-01  3.482447e-01  3.628087e-01  4.391634e-01
#> Flavniva  9.843366e-02  1.012485e-01  2.220446e-16  6.617758e-03  1.471559e-01
#> Nepharct  4.911667e-03  1.120851e-01  0.000000e+00  0.000000e+00  9.206808e-02
#> Stersp    1.657751e-01  2.429795e-01  2.853881e-03  2.902754e-02  3.384581e-01
#> Peltapht  9.110390e-02  2.643517e-01  1.004485e-01  9.644366e-02  2.753795e-01
#> Icmaeric  1.823830e-01  1.211264e-01  0.000000e+00  4.397555e-02  7.482820e-02
#> Cladcerv  7.654759e-02  1.109284e-01  0.000000e+00  0.000000e+00  7.664618e-02
#> Claddefo  2.663805e-01  5.014262e-01  8.458960e-02  1.725604e-01  5.484553e-01
#> Cladphyl  1.206318e-01  1.367853e-01  0.000000e+00  6.507592e-03  1.839799e-01
#>               Pinusylv      Descflex      Betupube      Vacculig      Diphcomp
#> Empenigr                                                                      
#> Rhodtome                                                                      
#> Vaccmyrt                                                                      
#> Vaccviti                                                                      
#> Pinusylv                                                                      
#> Descflex  1.146898e-01                                                        
#> Betupube  9.002433e-02  1.761084e-01                                          
#> Vacculig  7.581585e-02  2.323188e-01  7.225067e-02                            
#> Diphcomp  1.043300e-01  2.678571e-02  2.160494e-02  3.032397e-01              
#> Dicrsp    1.361918e-01  3.217813e-02  7.587910e-02  3.242095e-02  2.209877e-02
#> Dicrfusc  1.983933e-01  2.044473e-01  1.401333e-01  2.125629e-01  6.963694e-02
#> Dicrpoly  3.099660e-01  1.305398e-01  4.950495e-01  3.129242e-02  4.620462e-02
#> Hylosple  3.030573e-02  4.439341e-01  1.108647e-03  1.187171e-01  1.108647e-03
#> Pleuschr  3.434892e-01  3.258915e-01  1.432225e-01  2.054109e-01  5.972966e-02
#> Polypili  3.260341e-01  3.571429e-03  0.000000e+00  6.366731e-02  3.278689e-02
#> Polyjuni  2.125607e-01  6.140407e-02  8.701606e-02  4.207935e-02  6.853635e-02
#> Polycomm  1.192214e-01  2.938129e-01  4.210782e-01  7.687254e-02  2.160494e-02
#> Pohlnuta  6.224531e-01  1.832334e-01  1.335878e-01  1.643127e-01  1.611300e-01
#> Ptilcili  2.071826e-01  1.185531e-01  7.316203e-01  2.398465e-02  6.209850e-02
#> Barbhatc  9.531001e-02  1.179355e-01  8.620690e-01  2.507837e-02  2.160494e-02
#> Cladarbu  3.276037e-01  7.218122e-02  5.018624e-02  3.189986e-01  2.511352e-01
#> Cladrang  3.725850e-01  7.043683e-02  2.714106e-02  3.057116e-01  2.552295e-01
#> Cladstel  4.464561e-01  4.118238e-02  2.286783e-02  1.239621e-01  2.207683e-01
#> Cladunci  3.363980e-01  8.321784e-02  6.094527e-02  9.797035e-02  5.152807e-02
#> Cladcocc  4.627580e-01  1.073861e-01  1.039427e-01  2.059731e-01  2.422342e-01
#> Cladcorn  4.200796e-01  1.341410e-01  9.790442e-02  2.206903e-01  1.791612e-01
#> Cladgrac  5.239901e-01  1.599083e-01  1.167315e-01  2.535574e-01  2.377624e-01
#> Cladfimb  5.117415e-01  2.034271e-01  1.515152e-01  2.714995e-01  2.528058e-01
#> Cladcris  4.110359e-01  9.773379e-02  6.961178e-02  1.096739e-01  1.260226e-01
#> Cladchlo  3.459183e-01  1.190271e-01  2.327586e-01  1.120690e-01  1.892295e-01
#> Cladbotr  1.982710e-01  1.779635e-01  5.319149e-01  8.657422e-02  2.160494e-02
#> Cladamau  2.433090e-02  2.220446e-16  0.000000e+00  2.135348e-01  7.257496e-01
#> Cladsp    2.141119e-01  1.037088e-01  7.692308e-02  1.515213e-01  1.588319e-01
#> Cetreric  3.663423e-01  3.174603e-02  1.388889e-02  7.502555e-02  1.327160e-01
#> Cetrisla  5.735980e-01  1.179187e-01  3.399015e-01  1.510846e-02  9.686493e-02
#> Flavniva  1.173550e-01  1.954491e-02  0.000000e+00  4.416518e-02  9.873418e-02
#> Nepharct  1.247074e-02  2.678571e-02  6.896552e-02  7.742964e-02  9.194714e-02
#> Stersp    1.880187e-01  2.827789e-02  5.707763e-03  1.108498e-01  1.118721e-01
#> Peltapht  2.501601e-01  1.099624e-01  6.896552e-02  1.031883e-01  6.952567e-02
#> Icmaeric  5.596107e-02  8.928571e-03 -2.220446e-16  1.363636e-01  8.641975e-02
#> Cladcerv  2.433090e-02  2.678571e-02  6.896552e-02  5.256242e-03  2.160494e-02
#> Claddefo  3.920200e-01  1.014541e-01  5.865103e-02  1.153155e-01  9.296369e-02
#> Cladphyl  3.406326e-01  1.110223e-16  0.000000e+00  0.000000e+00  7.716049e-02
#>                 Dicrsp      Dicrfusc      Dicrpoly      Hylosple      Pleuschr
#> Empenigr                                                                      
#> Rhodtome                                                                      
#> Vaccmyrt                                                                      
#> Vaccviti                                                                      
#> Pinusylv                                                                      
#> Descflex                                                                      
#> Betupube                                                                      
#> Vacculig                                                                      
#> Diphcomp                                                                      
#> Dicrsp                                                                        
#> Dicrfusc  1.553396e-01                                                        
#> Dicrpoly  3.324948e-01  8.127586e-02                                          
#> Hylosple  2.950371e-02  5.413895e-02  4.620462e-02                            
#> Pleuschr  2.376412e-01  4.401227e-01  2.405864e-01  3.763501e-01              
#> Polypili  2.913580e-02  1.083683e-01  6.930693e-02  7.483370e-02  1.940825e-01
#> Polyjuni  3.884245e-01  2.109144e-01  1.194410e-01  1.049493e-02  3.009052e-01
#> Polycomm  1.892714e-01  1.604805e-01  3.554130e-01  1.842072e-01  3.111887e-01
#> Pohlnuta  1.974460e-01  2.360353e-01  4.340438e-01  8.015267e-02  4.643980e-01
#> Ptilcili  3.595844e-02  1.094844e-01  5.735967e-01  8.960182e-03  1.953613e-01
#> Barbhatc  6.666667e-03  5.831268e-02  5.016502e-01  2.507837e-02  5.751331e-02
#> Cladarbu  1.027008e-01  3.141440e-01  1.232209e-01  3.400416e-02  3.458751e-01
#> Cladrang  9.012860e-02  2.093078e-01  1.715536e-01  5.184056e-02  2.904998e-01
#> Cladstel  1.396954e-02  6.771130e-02  1.528070e-01  5.341990e-03  1.610561e-01
#> Cladunci  2.293293e-01  3.081825e-01  2.617130e-01  7.949601e-02  3.811130e-01
#> Cladcocc  9.549980e-02  3.942861e-01  1.994393e-01  1.902979e-02  3.946809e-01
#> Cladcorn  1.776769e-01  3.708702e-01  2.066367e-01  8.631765e-02  4.643694e-01
#> Cladgrac  2.017678e-01  3.289429e-01  3.091074e-01  8.087518e-02  5.352964e-01
#> Cladfimb  1.131874e-01  4.111757e-01  2.418492e-01  9.706824e-02  4.445545e-01
#> Cladcris  7.527592e-02  3.727995e-01  1.978475e-01  1.717290e-02  3.729354e-01
#> Cladchlo  9.361430e-02  9.174139e-02  3.957835e-01 -2.220446e-16  2.246050e-01
#> Cladbotr  7.761492e-02  1.102008e-01  5.904080e-01  1.276596e-01  2.778139e-01
#> Cladamau  2.220446e-16  2.131783e-02  4.950495e-03  0.000000e+00  1.256713e-02
#> Cladsp    8.586895e-02  2.127243e-01  9.240924e-02  7.803172e-02  2.320799e-01
#> Cetreric  2.805556e-01  2.751351e-01  3.635314e-01  4.277531e-02  2.868393e-01
#> Cetrisla  1.225458e-01  1.022428e-01  5.640313e-01  1.096086e-02  2.372928e-01
#> Flavniva  0.000000e+00  3.857593e-02  7.562769e-02  0.000000e+00  3.069924e-02
#> Nepharct  3.499751e-01  1.117272e-01  8.752776e-03  4.910929e-03  8.411857e-02
#> Stersp    1.891595e-02  1.437318e-01  7.989172e-02  2.250200e-03  2.123130e-01
#> Peltapht  9.259909e-02  1.927043e-01  4.455446e-02  9.321391e-02  2.933911e-01
#> Icmaeric  2.518519e-02  3.243481e-01  4.785479e-02 -2.220446e-16  7.106384e-02
#> Cladcerv  2.000000e-01  9.733968e-02  0.000000e+00  1.108647e-03  8.214938e-02
#> Claddefo  1.124543e-01  4.163910e-01  1.990118e-01  9.842524e-02  4.339433e-01
#> Cladphyl  0.000000e+00  7.025194e-02  1.237624e-01  0.000000e+00  7.987406e-02
#>               Polypili      Polyjuni      Polycomm      Pohlnuta      Ptilcili
#> Empenigr                                                                      
#> Rhodtome                                                                      
#> Vaccmyrt                                                                      
#> Vaccviti                                                                      
#> Pinusylv                                                                      
#> Descflex                                                                      
#> Betupube                                                                      
#> Vacculig                                                                      
#> Diphcomp                                                                      
#> Dicrsp                                                                        
#> Dicrfusc                                                                      
#> Dicrpoly                                                                      
#> Hylosple                                                                      
#> Pleuschr                                                                      
#> Polypili                                                                      
#> Polyjuni  1.659348e-01                                                        
#> Polycomm  3.278689e-02  3.368892e-01                                          
#> Pohlnuta  2.449631e-01  2.538540e-01  1.564885e-01                            
#> Ptilcili  1.155849e-01  2.424153e-01  3.927979e-01  2.881012e-01              
#> Barbhatc  1.253918e-02  3.203377e-02  3.834606e-01  1.270430e-01  7.498697e-01
#> Cladarbu  3.129987e-01  2.638837e-01  8.835329e-02  3.640181e-01  1.682634e-01
#> Cladrang  3.966491e-01  2.331953e-01  5.363896e-02  4.116334e-01  1.695491e-01
#> Cladstel  1.311206e-01  1.449612e-01  7.324690e-02  4.571788e-01  1.487323e-01
#> Cladunci  2.390390e-01  1.569516e-01  5.393305e-02  4.016029e-01  1.340704e-01
#> Cladcocc  3.590105e-01  2.995846e-01  1.070221e-01  5.711374e-01  1.764254e-01
#> Cladcorn  2.559169e-01  4.928634e-01  2.291336e-01  4.784247e-01  2.768268e-01
#> Cladgrac  3.159724e-01  3.740156e-01  1.682468e-01  6.616568e-01  3.036589e-01
#> Cladfimb  2.435006e-01  2.786056e-01  1.796842e-01  6.280554e-01  2.138408e-01
#> Cladcris  1.927491e-01  3.561645e-01  9.108735e-02  4.456452e-01  2.941015e-01
#> Cladchlo  9.143584e-02  1.844018e-01  2.586207e-01  4.938800e-01  3.272539e-01
#> Cladbotr  1.081270e-01  1.944235e-01  4.372191e-01  3.179308e-01  6.854071e-01
#> Cladamau  2.470726e-01  5.631769e-02  1.110223e-16  6.106870e-02  1.070664e-02
#> Cladsp    1.204288e-01  8.539295e-02  7.692308e-02  2.871403e-01  8.065667e-02
#> Cetreric  1.791894e-01  1.851785e-01  3.650235e-02  4.550254e-01  7.027718e-02
#> Cetrisla  1.379310e-01  1.391759e-01  3.497537e-01  6.357124e-01  4.282690e-01
#> Flavniva  4.388186e-02  1.035385e-01  1.687764e-03  1.331240e-01  3.763677e-02
#> Nepharct  7.604563e-03  5.294341e-01  1.830986e-01  3.050533e-02  2.164668e-02
#> Stersp    4.642189e-01  2.686770e-01  4.052511e-02  2.621562e-01  1.798202e-01
#> Peltapht  3.880500e-01  2.740547e-01  1.842105e-01  2.057051e-01  2.241256e-01
#> Icmaeric  2.950820e-01  8.808664e-02  2.816901e-02  8.778626e-02  7.137759e-03
#> Cladcerv  0.000000e+00  2.093863e-01  1.830986e-01  1.908397e-02  1.784440e-02
#> Claddefo  2.893611e-01  3.741858e-01  1.229882e-01  4.612836e-01  2.666863e-01
#> Cladphyl  3.125000e-01  7.581227e-02  2.816901e-02  3.129771e-01  6.566738e-02
#>               Barbhatc      Cladarbu      Cladrang      Cladstel      Cladunci
#> Empenigr                                                                      
#> Rhodtome                                                                      
#> Vaccmyrt                                                                      
#> Vaccviti                                                                      
#> Pinusylv                                                                      
#> Descflex                                                                      
#> Betupube                                                                      
#> Vacculig                                                                      
#> Diphcomp                                                                      
#> Dicrsp                                                                        
#> Dicrfusc                                                                      
#> Dicrpoly                                                                      
#> Hylosple                                                                      
#> Pleuschr                                                                      
#> Polypili                                                                      
#> Polyjuni                                                                      
#> Polycomm                                                                      
#> Pohlnuta                                                                      
#> Ptilcili                                                                      
#> Barbhatc                                                                      
#> Cladarbu  4.279485e-02                                                        
#> Cladrang  6.620306e-02  6.947622e-01                                          
#> Cladstel  5.595042e-02  2.485122e-01  4.150106e-01                            
#> Cladunci  3.469887e-02  4.046673e-01  3.150136e-01  1.589484e-01              
#> Cladcocc  3.448276e-02  6.593540e-01  5.905066e-01  4.036890e-01  3.915312e-01
#> Cladcorn  5.369473e-02  5.218092e-01  4.731120e-01  3.258893e-01  3.791938e-01
#> Cladgrac  1.064672e-01  6.138697e-01  6.005964e-01  3.472612e-01  4.748333e-01
#> Cladfimb  1.226924e-01  6.022942e-01  5.552486e-01  3.768210e-01  3.947578e-01
#> Cladcris  7.196603e-02  5.013560e-01  4.193193e-01  2.727539e-01  4.224146e-01
#> Cladchlo  2.452978e-01  2.446843e-01  2.684727e-01  5.053833e-01  1.911440e-01
#> Cladbotr  5.914760e-01  2.040776e-01  1.743716e-01  9.498688e-02  1.273949e-01
#> Cladamau  0.000000e+00  1.922368e-01  2.829101e-01  1.234411e-01  1.474769e-02
#> Cladsp    4.702194e-02  3.577104e-01  3.185047e-01  2.918576e-01  2.056886e-01
#> Cetreric  2.194357e-02  4.917571e-01  4.288223e-01  2.679058e-01  5.803542e-01
#> Cetrisla  3.399015e-01  1.732980e-01  2.699655e-01  4.924741e-01  2.132564e-01
#> Flavniva  3.375527e-03  1.445242e-01  1.588663e-01  1.665569e-01  5.294819e-02
#> Nepharct  0.000000e+00  9.766929e-02  7.993286e-02  3.299741e-02  3.248018e-02
#> Stersp    1.425151e-02  4.051092e-01  4.495809e-01  1.676475e-01  1.801134e-01
#> Peltapht  5.329154e-02  1.552389e-01  1.784323e-01  1.023197e-01  9.701493e-02
#> Icmaeric  0.000000e+00  3.052717e-01  3.006097e-01  2.876456e-03  1.166731e-01
#> Cladcerv  0.000000e+00  8.708096e-02  1.203468e-01  2.120359e-01  3.073916e-02
#> Claddefo  8.824620e-02  4.472966e-01  3.907133e-01  2.146155e-01  4.846953e-01
#> Cladphyl  0.000000e+00  1.073147e-01  1.127319e-01  4.119907e-01  1.163380e-01
#>               Cladcocc      Cladcorn      Cladgrac      Cladfimb      Cladcris
#> Empenigr                                                                      
#> Rhodtome                                                                      
#> Vaccmyrt                                                                      
#> Vaccviti                                                                      
#> Pinusylv                                                                      
#> Descflex                                                                      
#> Betupube                                                                      
#> Vacculig                                                                      
#> Diphcomp                                                                      
#> Dicrsp                                                                        
#> Dicrfusc                                                                      
#> Dicrpoly                                                                      
#> Hylosple                                                                      
#> Pleuschr                                                                      
#> Polypili                                                                      
#> Polyjuni                                                                      
#> Polycomm                                                                      
#> Pohlnuta                                                                      
#> Ptilcili                                                                      
#> Barbhatc                                                                      
#> Cladarbu                                                                      
#> Cladrang                                                                      
#> Cladstel                                                                      
#> Cladunci                                                                      
#> Cladcocc                                                                      
#> Cladcorn  5.760237e-01                                                        
#> Cladgrac  6.331255e-01  6.760982e-01                                          
#> Cladfimb  7.245031e-01  6.013999e-01  7.603860e-01                            
#> Cladcris  5.216133e-01  6.905349e-01  5.733830e-01  5.576549e-01              
#> Cladchlo  3.586083e-01  3.178567e-01  4.022877e-01  4.018635e-01  2.640908e-01
#> Cladbotr  1.433692e-01  2.614764e-01  3.477523e-01  2.698259e-01  2.576547e-01
#> Cladamau  1.756272e-01  1.141479e-01  1.420233e-01  1.262626e-01  5.087015e-02
#> Cladsp    4.086711e-01  2.666337e-01  3.741395e-01  4.085082e-01  2.637216e-01
#> Cetreric  5.677419e-01  3.977045e-01  5.079875e-01  5.239899e-01  4.342369e-01
#> Cetrisla  3.843600e-01  2.804635e-01  3.718924e-01  3.707519e-01  2.660758e-01
#> Flavniva  1.894288e-01  1.613551e-01  1.498793e-01  1.788390e-01  1.316523e-01
#> Nepharct  1.078063e-01  1.599159e-01  1.072184e-01  1.120233e-01  4.776521e-02
#> Stersp    3.877473e-01  3.667173e-01  4.021307e-01  2.998478e-01  3.233046e-01
#> Peltapht  2.139219e-01  4.555762e-01  2.549662e-01  1.586922e-01  3.377721e-01
#> Icmaeric  2.437276e-01  1.575563e-01  1.653696e-01  1.843434e-01  1.377632e-01
#> Cladcerv  7.885305e-02  1.205788e-01  6.420233e-02  7.575758e-02  3.346720e-02
#> Claddefo  4.776800e-01  6.305708e-01  6.034380e-01  5.949006e-01  7.435568e-01
#> Cladphyl  3.313172e-01  1.830788e-01  1.264591e-01  1.843434e-01  1.629016e-01
#>               Cladchlo      Cladbotr      Cladamau        Cladsp      Cetreric
#> Empenigr                                                                      
#> Rhodtome                                                                      
#> Vaccmyrt                                                                      
#> Vaccviti                                                                      
#> Pinusylv                                                                      
#> Descflex                                                                      
#> Betupube                                                                      
#> Vacculig                                                                      
#> Diphcomp                                                                      
#> Dicrsp                                                                        
#> Dicrfusc                                                                      
#> Dicrpoly                                                                      
#> Hylosple                                                                      
#> Pleuschr                                                                      
#> Polypili                                                                      
#> Polyjuni                                                                      
#> Polycomm                                                                      
#> Pohlnuta                                                                      
#> Ptilcili                                                                      
#> Barbhatc                                                                      
#> Cladarbu                                                                      
#> Cladrang                                                                      
#> Cladstel                                                                      
#> Cladunci                                                                      
#> Cladcocc                                                                      
#> Cladcorn                                                                      
#> Cladgrac                                                                      
#> Cladfimb                                                                      
#> Cladcris                                                                      
#> Cladchlo                                                                      
#> Cladbotr  3.609685e-01                                                        
#> Cladamau  2.586207e-02  2.220446e-16                                          
#> Cladsp    2.758621e-01  2.115385e-01  3.846154e-02                            
#> Cetreric  3.015326e-01  1.342199e-01  6.944444e-02  3.512821e-01              
#> Cetrisla  5.307882e-01  3.923069e-01  2.220446e-16  2.059492e-01  3.048577e-01
#> Flavniva  1.198312e-01  3.375527e-03  4.641350e-02  1.695716e-01  2.178973e-01
#> Nepharct  6.388488e-02  3.802281e-03  3.231939e-02  4.606610e-02  4.371567e-02
#> Stersp    1.475555e-01  1.107306e-01  2.656556e-01  1.004566e-01  2.248858e-01
#> Peltapht  6.034483e-02  1.752520e-01  2.631579e-02  1.295547e-01  4.020468e-02
#> Icmaeric  6.896552e-02  0.000000e+00  2.142857e-01  3.846154e-02  2.047980e-01
#> Cladcerv  2.586207e-02  0.000000e+00  1.110223e-16  1.346154e-01  8.611111e-02
#> Claddefo  2.778002e-01  3.248892e-01  5.669599e-02  2.411460e-01  5.071033e-01
#> Cladphyl  4.310345e-01  0.000000e+00  1.110223e-16  3.125000e-01  2.013889e-01
#>               Cetrisla      Flavniva      Nepharct        Stersp      Peltapht
#> Empenigr                                                                      
#> Rhodtome                                                                      
#> Vaccmyrt                                                                      
#> Vaccviti                                                                      
#> Pinusylv                                                                      
#> Descflex                                                                      
#> Betupube                                                                      
#> Vacculig                                                                      
#> Diphcomp                                                                      
#> Dicrsp                                                                        
#> Dicrfusc                                                                      
#> Dicrpoly                                                                      
#> Hylosple                                                                      
#> Pleuschr                                                                      
#> Polypili                                                                      
#> Polyjuni                                                                      
#> Polycomm                                                                      
#> Pohlnuta                                                                      
#> Ptilcili                                                                      
#> Barbhatc                                                                      
#> Cladarbu                                                                      
#> Cladrang                                                                      
#> Cladstel                                                                      
#> Cladunci                                                                      
#> Cladcocc                                                                      
#> Cladcorn                                                                      
#> Cladgrac                                                                      
#> Cladfimb                                                                      
#> Cladcris                                                                      
#> Cladchlo                                                                      
#> Cladbotr                                                                      
#> Cladamau                                                                      
#> Cladsp                                                                        
#> Cetreric                                                                      
#> Cetrisla                                                                      
#> Flavniva  1.022677e-01                                                        
#> Nepharct  9.852217e-03  3.586979e-02                                          
#> Stersp    1.285990e-01  1.394038e-01  6.351459e-02                            
#> Peltapht  1.428571e-01  1.181435e-02  9.590754e-02  1.770007e-01              
#> Icmaeric  4.926108e-02  4.303797e-02  3.802281e-02  5.281756e-01  0.000000e+00
#> Cladcerv  3.448276e-02  5.000000e-01  2.000000e-01  5.993151e-02  9.210526e-02
#> Claddefo  2.390920e-01  1.148323e-01  3.008746e-02  2.749551e-01  2.628749e-01
#> Cladphyl  3.793103e-01  7.763713e-02  0.000000e+00  8.618721e-02  3.125000e-01
#>               Icmaeric      Cladcerv      Claddefo
#> Empenigr                                          
#> Rhodtome                                          
#> Vaccmyrt                                          
#> Vaccviti                                          
#> Pinusylv                                          
#> Descflex                                          
#> Betupube                                          
#> Vacculig                                          
#> Diphcomp                                          
#> Dicrsp                                            
#> Dicrfusc                                          
#> Dicrpoly                                          
#> Hylosple                                          
#> Pleuschr                                          
#> Polypili                                          
#> Polyjuni                                          
#> Polycomm                                          
#> Pohlnuta                                          
#> Ptilcili                                          
#> Barbhatc                                          
#> Cladarbu                                          
#> Cladrang                                          
#> Cladstel                                          
#> Cladunci                                          
#> Cladcocc                                          
#> Cladcorn                                          
#> Cladgrac                                          
#> Cladfimb                                          
#> Cladcris                                          
#> Cladchlo                                          
#> Cladbotr                                          
#> Cladamau                                          
#> Cladsp                                            
#> Cetreric                                          
#> Cetrisla                                          
#> Flavniva                                          
#> Nepharct                                          
#> Stersp                                            
#> Peltapht                                          
#> Icmaeric                                          
#> Cladcerv  0.000000e+00                            
#> Claddefo  1.906158e-01  1.564027e-02              
#> Cladphyl  6.250000e-02  0.000000e+00  1.113759e-01